Reorganize files in the DataFrame related projects
