+++ src/Microsoft.Data.Analysis/DataFrameColumns/VBufferDataFrameColumn.cs
