+++ src/Microsoft.Data.Analysis/Utils/BitUtility.cs
