+++ test/Microsoft.Data.Analysis.Tests/DataFrameTests.cs
@@ -6,276 +6,14 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using Apache.Arrow;
 using Microsoft.ML;
 using Microsoft.ML.Data;
-using Microsoft.ML.TestFramework.Attributes;
 using Xunit;
 
 namespace Microsoft.Data.Analysis.Tests
 {
     public partial class DataFrameTests
     {
-        public static DataFrame MakeDataFrameWithTwoColumns(int length, bool withNulls = true)
-        {
-            DataFrameColumn dataFrameColumn1 = new Int32DataFrameColumn("Int1", Enumerable.Range(0, length).Select(x => x));
-            DataFrameColumn dataFrameColumn2 = new Int32DataFrameColumn("Int2", Enumerable.Range(10, length).Select(x => x));
-            if (withNulls)
-            {
-                dataFrameColumn1[length / 2] = null;
-                dataFrameColumn2[length / 2] = null;
-            }
-            DataFrame dataFrame = new DataFrame();
-            dataFrame.Columns.Insert(0, dataFrameColumn1);
-            dataFrame.Columns.Insert(1, dataFrameColumn2);
-            return dataFrame;
-        }
-
-        public static ArrowStringDataFrameColumn CreateArrowStringColumn(int length, bool withNulls = true)
-        {
-            byte[] dataMemory = new byte[length * 3];
-            byte[] nullMemory = new byte[BitUtility.ByteCount(length)];
-            byte[] offsetMemory = new byte[(length + 1) * 4];
-
-            // Initialize offset with 0 as the first value
-            offsetMemory[0] = 0;
-            offsetMemory[1] = 0;
-            offsetMemory[2] = 0;
-            offsetMemory[3] = 0;
-
-            // Append "foo" length times, with a possible `null` in the middle
-            int validStringsIndex = 0;
-            for (int i = 0; i < length; i++)
-            {
-                if (withNulls && i == length / 2)
-                {
-                    BitUtility.SetBit(nullMemory, i, false);
-                }
-                else
-                {
-                    int dataMemoryIndex = validStringsIndex * 3;
-                    dataMemory[dataMemoryIndex++] = 102;
-                    dataMemory[dataMemoryIndex++] = 111;
-                    dataMemory[dataMemoryIndex++] = 111;
-                    BitUtility.SetBit(nullMemory, i, true);
-
-                    validStringsIndex++;
-                }
-
-                // write the current length to (index + 1)
-                int offsetIndex = (i + 1) * 4;
-                int offsetValue = 3 * validStringsIndex;
-                byte[] offsetValueBytes = BitConverter.GetBytes(offsetValue);
-                offsetMemory[offsetIndex++] = offsetValueBytes[0];
-                offsetMemory[offsetIndex++] = offsetValueBytes[1];
-                offsetMemory[offsetIndex++] = offsetValueBytes[2];
-                offsetMemory[offsetIndex++] = offsetValueBytes[3];
-            }
-
-            int nullCount = withNulls ? 1 : 0;
-            return new ArrowStringDataFrameColumn("ArrowString", dataMemory, offsetMemory, nullMemory, length, nullCount);
-        }
-
-        public static VBufferDataFrameColumn<int> CreateVBufferDataFrameColumn(int length)
-        {
-            var buffers = Enumerable.Repeat(new VBuffer<int>(5, new[] { 0, 1, 2, 3, 4 }), length).ToArray();
-            return new VBufferDataFrameColumn<int>("VBuffer", buffers);
-        }
-
-        public static DataFrame MakeDataFrameWithAllColumnTypes(int length, bool withNulls = true)
-        {
-            DataFrame df = MakeDataFrameWithAllMutableAndArrowColumnTypes(length, withNulls);
-
-            var vBufferColumn = CreateVBufferDataFrameColumn(length);
-            df.Columns.Insert(df.Columns.Count, vBufferColumn);
-
-            return df;
-        }
-
-        public static DataFrame MakeDataFrameWithAllMutableAndArrowColumnTypes(int length, bool withNulls = true)
-        {
-            DataFrame df = MakeDataFrameWithAllMutableColumnTypes(length, withNulls);
-            DataFrameColumn arrowStringColumn = CreateArrowStringColumn(length, withNulls);
-            df.Columns.Insert(df.Columns.Count, arrowStringColumn);
-
-            return df;
-        }
-
-        public static DataFrame MakeDataFrameWithAllMutableColumnTypes(int length, bool withNulls = true)
-        {
-            DataFrame df = MakeDataFrameWithNumericStringAndDateTimeColumns(length, withNulls);
-            DataFrameColumn boolColumn = new BooleanDataFrameColumn("Bool", Enumerable.Range(0, length).Select(x => x % 2 == 0));
-            df.Columns.Insert(df.Columns.Count, boolColumn);
-            if (withNulls)
-            {
-                boolColumn[length / 2] = null;
-            }
-            return df;
-        }
-
-        public static DataFrame MakeDataFrameWithNumericAndBoolColumns(int length, bool withNulls = true)
-        {
-            DataFrame df = MakeDataFrameWithNumericColumns(length, withNulls);
-            DataFrameColumn boolColumn = new BooleanDataFrameColumn("Bool", Enumerable.Range(0, length).Select(x => x % 2 == 0));
-            df.Columns.Insert(df.Columns.Count, boolColumn);
-            if (withNulls)
-            {
-                boolColumn[length / 2] = null;
-            }
-            return df;
-        }
-
-        public static DataFrame MakeDataFrameWithNumericAndStringColumns(int length, bool withNulls = true)
-        {
-            DataFrame df = MakeDataFrameWithNumericColumns(length, withNulls);
-            DataFrameColumn stringColumn = new StringDataFrameColumn("String", Enumerable.Range(0, length).Select(x => x.ToString()));
-            df.Columns.Insert(df.Columns.Count, stringColumn);
-            if (withNulls)
-            {
-                stringColumn[length / 2] = null;
-            }
-
-            DataFrameColumn charColumn = new CharDataFrameColumn("Char", Enumerable.Range(0, length).Select(x => (char)(x + 65)));
-            df.Columns.Insert(df.Columns.Count, charColumn);
-            if (withNulls)
-            {
-                charColumn[length / 2] = null;
-            }
-            return df;
-        }
-
-        internal static DateTime SampleDateTime = new DateTime(2021, 06, 04);
-        public static DataFrame MakeDataFrameWithNumericStringAndDateTimeColumns(int length, bool withNulls = true)
-        {
-            DataFrame df = MakeDataFrameWithNumericAndStringColumns(length, withNulls);
-
-            DataFrameColumn dateTimeColumn = new DateTimeDataFrameColumn("DateTime", Enumerable.Range(0, length).Select(x => SampleDateTime.AddDays(x)));
-            df.Columns.Insert(df.Columns.Count, dateTimeColumn);
-            if (withNulls)
-            {
-                dateTimeColumn[length / 2] = null;
-            }
-            return df;
-        }
-
-        public static DataFrame MakeDataFrameWithNumericColumns(int length, bool withNulls = true, int startingFrom = 0)
-        {
-            IEnumerable<int> range = Enumerable.Range(startingFrom, length);
-
-            var byteColumn = new ByteDataFrameColumn("Byte", range.Select(x => (byte)x));
-            var decimalColumn = new DecimalDataFrameColumn("Decimal", range.Select(x => (decimal)x));
-            var doubleColumn = new DoubleDataFrameColumn("Double", range.Select(x => (double)x));
-            var floatColumn = new SingleDataFrameColumn("Float", range.Select(x => (float)x));
-            var intColumn = new Int32DataFrameColumn("Int", range.Select(x => x));
-            var longColumn = new Int64DataFrameColumn("Long", range.Select(x => (long)x));
-            var sbyteColumn = new SByteDataFrameColumn("Sbyte", range.Select(x => (sbyte)x));
-            var shortColumn = new Int16DataFrameColumn("Short", range.Select(x => (short)x));
-            var uintColumn = new UInt32DataFrameColumn("Uint", range.Select(x => (uint)x));
-            var ulongColumn = new UInt64DataFrameColumn("Ulong", range.Select(x => (ulong)x));
-            var ushortColumn = new UInt16DataFrameColumn("Ushort", range.Select(x => (ushort)x));
-
-            var columnsList = new List<DataFrameColumn>
-            {
-                byteColumn,
-                decimalColumn,
-                doubleColumn,
-                floatColumn,
-                intColumn,
-                longColumn,
-                sbyteColumn,
-                shortColumn,
-                uintColumn,
-                ulongColumn,
-                ushortColumn
-            };
-
-            var dataFrame = new DataFrame(columnsList);
-
-            if (withNulls)
-            {
-                for (var i = 0; i < dataFrame.Columns.Count; i++)
-                {
-                    dataFrame.Columns[i][length / 2] = null;
-                }
-            }
-
-            return dataFrame;
-        }
-
-        public static DataFrame MakeDataFrame<T1, T2>(int length, bool withNulls = true)
-            where T1 : unmanaged
-            where T2 : unmanaged
-        {
-            DataFrameColumn baseColumn1 = DataFrameColumn.Create("Column1", Enumerable.Range(0, length).Select(x => (T1)Convert.ChangeType(x % 2 == 0 ? 0 : 1, typeof(T1))));
-            DataFrameColumn baseColumn2 = DataFrameColumn.Create("Column2", Enumerable.Range(0, length).Select(x => (T2)Convert.ChangeType(x % 2 == 0 ? 0 : 1, typeof(T2))));
-            DataFrame dataFrame = new DataFrame(new List<DataFrameColumn> { baseColumn1, baseColumn2 });
-
-            if (withNulls)
-            {
-                for (int i = 0; i < dataFrame.Columns.Count; i++)
-                {
-                    dataFrame.Columns[i][length / 2] = null;
-                }
-            }
-
-            return dataFrame;
-        }
-
-        public DataFrame SplitTrainTest(DataFrame input, float testRatio, out DataFrame Test)
-        {
-            IEnumerable<int> randomIndices = Enumerable.Range(0, (int)input.Rows.Count);
-            IEnumerable<int> trainIndices = randomIndices.Take((int)(input.Rows.Count * testRatio));
-            IEnumerable<int> testIndices = randomIndices.Skip((int)(input.Rows.Count * testRatio));
-            Test = input[testIndices];
-            return input[trainIndices];
-        }
-
-        [Fact]
-        public void TestVBufferColumn_Creation()
-        {
-            var vBufferColumn = CreateVBufferDataFrameColumn(10);
-
-            Assert.Equal(10, vBufferColumn.Length);
-            Assert.Equal(5, vBufferColumn[0].GetValues().Length);
-            Assert.Equal(0, vBufferColumn[0].GetValues()[0]);
-        }
-
-        [Fact]
-        public void TestVBufferColumn_Indexer()
-        {
-            var buffer = new VBuffer<int>(5, new[] { 4, 3, 2, 1, 0 });
-
-            var vBufferColumn = new VBufferDataFrameColumn<int>("VBuffer", 1);
-            vBufferColumn[0] = buffer;
-
-            Assert.Equal(1, vBufferColumn.Length);
-            Assert.Equal(5, vBufferColumn[0].GetValues().Length);
-            Assert.Equal(0, vBufferColumn[0].GetValues()[4]);
-        }
-
-        [X64Fact("32-bit doesn't allow to allocate more than 2 Gb")]
-        public void TestVBufferColumn_Indexer_MoreThanMaxInt()
-        {
-            var originalValues = new[] { 4, 3, 2, 1, 0 };
-
-            var length = VBufferDataFrameColumn<int>.MaxCapacity + 3;
-
-            var vBufferColumn = new VBufferDataFrameColumn<int>("VBuffer", length);
-            long index = length - 2;
-
-            vBufferColumn[index] = new VBuffer<int>(5, originalValues);
-
-            var values = vBufferColumn[index].GetValues();
-
-            Assert.Equal(length, vBufferColumn.Length);
-            Assert.Equal(5, values.Length);
-
-            for (int i = 0; i < values.Length; i++)
-            {
-                Assert.Equal(originalValues[i], values[i]);
-            }
-        }
-
         [Fact]
         public void TestIndexer()
         {
@@ -464,418 +202,6 @@ namespace Microsoft.Data.Analysis.Tests
         }
 
         [Fact]
-        public void TestBinaryOperations()
-        {
-            DataFrame df = MakeDataFrameWithTwoColumns(12);
-            IReadOnlyList<int> listOfInts = new List<int>() { 5, 5 };
-
-            // The following binary ops return a copy
-            var ret = df.Add(5);
-            Assert.Equal(0, df[0, 0]);
-            Assert.Equal(5, ret[0, 0]);
-            ret = df.Add(listOfInts);
-            Assert.Equal(0, df[0, 0]);
-            Assert.Equal(5, ret[0, 0]);
-            ret = df.Subtract(5);
-            Assert.Equal(0, df[0, 0]);
-            Assert.Equal(-5, ret[0, 0]);
-            ret = df.Subtract(listOfInts);
-            Assert.Equal(0, df[0, 0]);
-            Assert.Equal(-5, ret[0, 0]);
-            ret = df.Multiply(5);
-            Assert.Equal(1, df[1, 0]);
-            Assert.Equal(5, ret[1, 0]);
-            ret = df.Multiply(listOfInts);
-            Assert.Equal(1, df[1, 0]);
-            Assert.Equal(5, ret[1, 0]);
-            ret = df.Divide(5);
-            Assert.Equal(5, df[5, 0]);
-            Assert.Equal(1, ret[5, 0]);
-            ret = df.Divide(listOfInts);
-            Assert.Equal(5, df[5, 0]);
-            Assert.Equal(1, ret[5, 0]);
-            ret = df.Modulo(5);
-            Assert.Equal(5, df[5, 0]);
-            Assert.Equal(0, ret[5, 0]);
-            ret = df.Modulo(listOfInts);
-            Assert.Equal(5, df[5, 0]);
-            Assert.Equal(0, ret[5, 0]);
-
-            Assert.Equal(true, df.ElementwiseGreaterThanOrEqual(5)[7, 0]);
-            Assert.Equal(true, df.ElementwiseGreaterThanOrEqual(listOfInts)[7, 0]);
-            Assert.Equal(true, df.ElementwiseLessThanOrEqual(5)[4, 0]);
-            Assert.Equal(true, df.ElementwiseLessThanOrEqual(listOfInts)[4, 0]);
-            Assert.Equal(false, df.ElementwiseGreaterThan(5)[5, 0]);
-            Assert.Equal(false, df.ElementwiseGreaterThan(listOfInts)[5, 0]);
-            Assert.Equal(false, df.ElementwiseLessThan(5)[5, 0]);
-            Assert.Equal(false, df.ElementwiseLessThan(listOfInts)[5, 0]);
-            // The following binary ops are in place
-            Assert.Equal(5, df.Add(5, inPlace: true)[0, 0]);
-            Assert.Equal(10, df.Add(listOfInts, inPlace: true)[0, 0]);
-            Assert.Equal(5, df.Subtract(5, inPlace: true)[0, 0]);
-            Assert.Equal(0, df.Subtract(listOfInts, inPlace: true)[0, 0]);
-            Assert.Equal(5, df.Multiply(5, inPlace: true)[1, 0]);
-            Assert.Equal(25, df.Multiply(listOfInts, inPlace: true)[1, 0]);
-            Assert.Equal(5, df.Divide(5, inPlace: true)[1, 0]);
-            Assert.Equal(1, df.Divide(listOfInts, inPlace: true)[1, 0]);
-            Assert.Equal(1, df.Modulo(5, inPlace: true)[1, 0]);
-            Assert.Equal(1, df.Modulo(listOfInts, inPlace: true)[1, 0]);
-            Assert.Equal(2, df.LeftShift(1)[1, 0]);
-            Assert.Equal(1, df.RightShift(1)[2, 0]);
-        }
-
-        [Fact]
-        public void TestBinaryOperationsWithColumns()
-        {
-            int length = 10;
-            var df1 = MakeDataFrameWithNumericColumns(length);
-            var df2 = MakeDataFrameWithNumericColumns(length);
-
-            DataFrameColumn newColumn;
-            DataFrameColumn verify;
-            for (int i = 0; i < df1.Columns.Count; i++)
-            {
-                newColumn = df1.Columns[df1.Columns[i].Name] + df2.Columns[df2.Columns[i].Name];
-                verify = newColumn.ElementwiseEquals(df1.Columns[i] * 2);
-                Assert.Equal(true, verify[0]);
-
-                newColumn = df1.Columns[df1.Columns[i].Name] - df2.Columns[df2.Columns[i].Name];
-                verify = newColumn.ElementwiseEquals(0);
-                Assert.Equal(true, verify[0]);
-
-                newColumn = df1.Columns[df1.Columns[i].Name] * df2.Columns[df2.Columns[i].Name];
-                verify = newColumn.ElementwiseEquals(df1.Columns[i] * df1.Columns[i]);
-                Assert.Equal(true, verify[0]);
-
-                var df1Column = df1.Columns[i] + 1;
-                var df2Column = df2.Columns[i] + 1;
-                newColumn = df1Column / df2Column;
-                verify = newColumn.ElementwiseEquals(1);
-                Assert.Equal(true, verify[0]);
-
-                newColumn = df1Column % df2Column;
-                verify = newColumn.ElementwiseEquals(0);
-                Assert.Equal(true, verify[0]);
-
-                verify = df1.Columns[df1.Columns[i].Name].ElementwiseEquals(df2.Columns[df2.Columns[i].Name]);
-                Assert.True(verify.All());
-
-                verify = df1.Columns[df1.Columns[i].Name].ElementwiseNotEquals(df2.Columns[df2.Columns[i].Name]);
-                Assert.False(verify.Any());
-
-                verify = df1.Columns[df1.Columns[i].Name].ElementwiseGreaterThanOrEqual(df2.Columns[df2.Columns[i].Name]);
-                Assert.True(verify.All());
-
-                verify = df1.Columns[df1.Columns[i].Name].ElementwiseLessThanOrEqual(df2.Columns[df2.Columns[i].Name]);
-                Assert.True(verify.All());
-
-                verify = df1.Columns[df1.Columns[i].Name].ElementwiseGreaterThan(df2.Columns[df2.Columns[i].Name]);
-                Assert.False(verify.Any());
-
-                verify = df1.Columns[df1.Columns[i].Name].ElementwiseLessThan(df2.Columns[df2.Columns[i].Name]);
-                Assert.False(verify.Any());
-            }
-        }
-
-        [Fact]
-        public void TestBinaryOperationsWithConversions()
-        {
-            DataFrame df = DataFrameTests.MakeDataFrameWithTwoColumns(10);
-
-            // Add a double to an int column
-            DataFrame dfd = df.Add(5.0f);
-            var dtype = dfd.Columns[0].DataType;
-            Assert.True(dtype == typeof(double));
-
-            // Add a decimal to an int column
-            DataFrame dfm = df.Add(5.0m);
-            dtype = dfm.Columns[0].DataType;
-            Assert.True(dtype == typeof(decimal));
-
-            // int + bool should throw
-            Assert.Throws<NotSupportedException>(() => df.Add(true));
-
-            var dataFrameColumn1 = new DoubleDataFrameColumn("Double1", Enumerable.Range(0, 10).Select(x => (double)x));
-            df.Columns[0] = dataFrameColumn1;
-            // Double + comparison ops should throw
-            Assert.Throws<NotSupportedException>(() => df.And(true));
-        }
-
-        [Fact]
-        public void TestBinaryOperationsOnBoolColumn()
-        {
-            var df = new DataFrame();
-            var dataFrameColumn1 = new BooleanDataFrameColumn("Bool1", Enumerable.Range(0, 10).Select(x => true));
-            var dataFrameColumn2 = new BooleanDataFrameColumn("Bool2", Enumerable.Range(0, 10).Select(x => true));
-            df.Columns.Insert(0, dataFrameColumn1);
-            df.Columns.Insert(1, dataFrameColumn2);
-
-            // bool + int should throw
-            Assert.Throws<NotSupportedException>(() => df.Add(5));
-            // Left shift should throw
-            Assert.Throws<NotSupportedException>(() => df.LeftShift(5));
-
-            IReadOnlyList<bool> listOfBools = new List<bool>() { true, false };
-            // boolean and And should work
-            var newdf = df.And(true);
-            Assert.Equal(true, newdf[4, 0]);
-            var newdf1 = df.And(listOfBools);
-            Assert.Equal(false, newdf1[4, 1]);
-
-            newdf = df.Or(true);
-            Assert.Equal(true, newdf[4, 0]);
-            newdf1 = df.Or(listOfBools);
-            Assert.Equal(true, newdf1[4, 1]);
-
-            newdf = df.Xor(true);
-            Assert.Equal(false, newdf[4, 0]);
-            newdf1 = df.Xor(listOfBools);
-            Assert.Equal(true, newdf1[4, 1]);
-        }
-
-        [Fact]
-        public void TestBinaryOperationsOnDateTimeColumn()
-        {
-            var df = new DataFrame();
-            var dataFrameColumn1 = new DateTimeDataFrameColumn("DateTime1", Enumerable.Range(0, 5).Select(x => SampleDateTime.AddDays(x)));
-            // Make the second data frame column have one value that is different
-            var dataFrameColumn2 = new DateTimeDataFrameColumn("DateTime2", Enumerable.Range(0, 4).Select(x => SampleDateTime.AddDays(x)));
-            dataFrameColumn2.Append(SampleDateTime.AddDays(6));
-            df.Columns.Insert(0, dataFrameColumn1);
-            df.Columns.Insert(1, dataFrameColumn2);
-
-            // DateTime + int should throw
-            Assert.Throws<NotSupportedException>(() => df.Add(5));
-            // Left shift should throw
-            Assert.Throws<NotSupportedException>(() => df.LeftShift(5));
-            // Right shift should throw
-            Assert.Throws<NotSupportedException>(() => df.RightShift(5));
-
-            // And should throw
-            Assert.Throws<NotSupportedException>(() => df.And(true));
-            // Or should throw
-            Assert.Throws<NotSupportedException>(() => df.Or(true));
-            // Xor should throw
-            Assert.Throws<NotSupportedException>(() => df.Xor(true));
-
-            var equalsResult = dataFrameColumn1.ElementwiseEquals(dataFrameColumn2);
-            Assert.True(equalsResult[0]);
-            Assert.False(equalsResult[4]);
-
-            var equalsToScalarResult = df["DateTime1"].ElementwiseEquals(SampleDateTime);
-            Assert.True(equalsToScalarResult[0]);
-            Assert.False(equalsToScalarResult[1]);
-
-            var notEqualsResult = dataFrameColumn1.ElementwiseNotEquals(dataFrameColumn2);
-            Assert.False(notEqualsResult[0]);
-            Assert.True(notEqualsResult[4]);
-
-            var notEqualsToScalarResult = df["DateTime1"].ElementwiseNotEquals(SampleDateTime);
-            Assert.False(notEqualsToScalarResult[0]);
-            Assert.True(notEqualsToScalarResult[1]);
-        }
-
-        [Fact]
-        public void TestBinaryOperationsOnArrowStringColumn()
-        {
-            var df = new DataFrame();
-            var strArrayBuilder = new StringArray.Builder();
-            for (int i = 0; i < 10; i++)
-            {
-                strArrayBuilder.Append(i.ToString());
-            }
-            StringArray strArray = strArrayBuilder.Build();
-
-            ArrowStringDataFrameColumn stringColumn = new ArrowStringDataFrameColumn("String", strArray.ValueBuffer.Memory, strArray.ValueOffsetsBuffer.Memory, strArray.NullBitmapBuffer.Memory, strArray.Length, strArray.NullCount);
-            df.Columns.Insert(0, stringColumn);
-
-            DataFrameColumn newCol = stringColumn.ElementwiseEquals(4);
-            Assert.Equal(true, newCol[4]);
-            Assert.Equal(false, newCol[0]);
-            Assert.Equal(false, newCol[5]);
-
-            newCol = stringColumn.ElementwiseEquals("4");
-            Assert.Equal(true, newCol[4]);
-            Assert.Equal(false, newCol[0]);
-
-            newCol = stringColumn.ElementwiseEquals("foo");
-            Assert.False(newCol.All());
-            newCol = stringColumn.ElementwiseEquals(null);
-            Assert.False(newCol.All());
-
-            ArrowStringDataFrameColumn stringColumnCopy = new ArrowStringDataFrameColumn("String", strArray.ValueBuffer.Memory, strArray.ValueOffsetsBuffer.Memory, strArray.NullBitmapBuffer.Memory, strArray.Length, strArray.NullCount);
-            newCol = stringColumn.ElementwiseEquals(stringColumnCopy);
-            Assert.True(newCol.All());
-
-            DataFrameColumn stringColumnCopyAsBaseColumn = stringColumnCopy;
-            newCol = stringColumn.ElementwiseEquals(stringColumnCopyAsBaseColumn);
-            Assert.True(newCol.All());
-
-            newCol = stringColumn.ElementwiseNotEquals(5);
-            Assert.Equal(true, newCol[0]);
-            Assert.Equal(false, newCol[5]);
-
-            newCol = stringColumn.ElementwiseNotEquals("5");
-            Assert.Equal(true, newCol[0]);
-            Assert.Equal(false, newCol[5]);
-
-            newCol = stringColumn.ElementwiseNotEquals("foo");
-            Assert.True(newCol.All());
-            newCol = stringColumn.ElementwiseNotEquals(null);
-            Assert.True(newCol.All());
-
-            newCol = stringColumn.ElementwiseNotEquals(stringColumnCopy);
-            Assert.False(newCol.All());
-
-            newCol = stringColumn.ElementwiseNotEquals(stringColumnCopyAsBaseColumn);
-            Assert.False(newCol.All());
-        }
-
-        [Fact]
-        public void TestBinaryOperationsOnStringColumn()
-        {
-            var df = new DataFrame();
-            DataFrameColumn stringColumn = new StringDataFrameColumn("String", Enumerable.Range(0, 10).Select(x => x.ToString()));
-            df.Columns.Insert(0, stringColumn);
-
-            DataFrameColumn newCol = stringColumn.ElementwiseEquals(5);
-            Assert.Equal(true, newCol[5]);
-            Assert.Equal(false, newCol[0]);
-
-            newCol = (stringColumn as StringDataFrameColumn).ElementwiseEquals("5");
-            Assert.Equal(true, newCol[5]);
-            Assert.Equal(false, newCol[0]);
-
-            DataFrameColumn stringColumnCopy = new StringDataFrameColumn("String", Enumerable.Range(0, 10).Select(x => x.ToString()));
-            newCol = stringColumn.ElementwiseEquals(stringColumnCopy);
-            Assert.Equal(true, newCol[5]);
-            Assert.Equal(true, newCol[0]);
-
-            StringDataFrameColumn typedStringColumn = stringColumn as StringDataFrameColumn;
-            StringDataFrameColumn typedStringColumnCopy = stringColumnCopy as StringDataFrameColumn;
-            newCol = typedStringColumn.ElementwiseEquals(typedStringColumnCopy);
-            Assert.True(newCol.All());
-
-            newCol = stringColumn.ElementwiseNotEquals(5);
-            Assert.Equal(false, newCol[5]);
-            Assert.Equal(true, newCol[0]);
-
-            newCol = typedStringColumn.ElementwiseNotEquals("5");
-            Assert.Equal(false, newCol[5]);
-            Assert.Equal(true, newCol[0]);
-
-            newCol = stringColumn.ElementwiseNotEquals(stringColumnCopy);
-            Assert.Equal(false, newCol[5]);
-            Assert.Equal(false, newCol[0]);
-
-            newCol = typedStringColumn.ElementwiseNotEquals(typedStringColumnCopy);
-            Assert.False(newCol.All());
-
-            newCol = typedStringColumn.Add("suffix");
-            for (int i = 0; i < newCol.Length; i++)
-            {
-                Assert.Equal(newCol[i], typedStringColumn[i] + "suffix");
-            }
-            DataFrameColumn addString = typedStringColumn + "suffix";
-            for (int i = 0; i < addString.Length; i++)
-            {
-                Assert.Equal(addString[i], typedStringColumn[i] + "suffix");
-            }
-            Assert.True(newCol.ElementwiseEquals(addString).All());
-            addString = "prefix" + typedStringColumn;
-            for (int i = 0; i < addString.Length; i++)
-            {
-                Assert.Equal(addString[i], "prefix" + typedStringColumn[i]);
-            }
-        }
-
-        [Fact]
-        public void TestBinaryOperatorsWithConversions()
-        {
-            var df = MakeDataFrameWithNumericColumns(10);
-
-            DataFrame tempDf = df + 1;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] + (double)1);
-            tempDf = df + 1.1;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] + 1.1);
-            tempDf = df + 1.1m;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] + 1.1m);
-            Assert.True(typeof(decimal) == tempDf.Columns["Int"].DataType);
-
-            tempDf = df - 1.1;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] - 1.1);
-            tempDf = df - 1.1m;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] - 1.1m);
-            Assert.True(typeof(decimal) == tempDf.Columns["Int"].DataType);
-
-            tempDf = df * 1.1;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] * 1.1);
-            tempDf = df * 1.1m;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] * 1.1m);
-            Assert.True(typeof(decimal) == tempDf.Columns["Int"].DataType);
-
-            tempDf = df / 1.1;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] / 1.1);
-            tempDf = df / 1.1m;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] / 1.1m);
-            Assert.True(typeof(decimal) == tempDf.Columns["Int"].DataType);
-
-            tempDf = df % 1.1;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] % 1.1);
-            tempDf = df % 1.1m;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] % 1.1m);
-            Assert.True(typeof(decimal) == tempDf.Columns["Int"].DataType);
-
-            tempDf = 1 + df;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] + (double)1);
-            tempDf = 1.1 + df;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] + 1.1);
-            tempDf = 1.1m + df;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] + 1.1m);
-            Assert.True(typeof(decimal) == tempDf.Columns["Int"].DataType);
-
-            tempDf = 1.1 - df;
-            Assert.Equal(tempDf[0, 0], 1.1 - (byte)df[0, 0]);
-            tempDf = 1.1m - df;
-            Assert.Equal(tempDf[0, 0], 1.1m - (byte)df[0, 0]);
-            Assert.True(typeof(decimal) == tempDf.Columns["Int"].DataType);
-
-            tempDf = 1.1 * df;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] * 1.1);
-            tempDf = 1.1m * df;
-            Assert.Equal(tempDf[0, 0], (byte)df[0, 0] * 1.1m);
-            Assert.True(typeof(decimal) == tempDf.Columns["Int"].DataType);
-
-            // To prevent a divide by zero
-            var plusOne = df + 1;
-            tempDf = 1.1 / plusOne;
-            Assert.Equal(tempDf[0, 0], 1.1 / (double)plusOne[0, 0]);
-            var plusDecimal = df + 1.1m;
-            tempDf = 1.1m / plusDecimal;
-            Assert.Equal(tempDf[0, 0], (1.1m) / (decimal)plusDecimal[0, 0]);
-            Assert.True(typeof(decimal) == tempDf.Columns["Int"].DataType);
-
-            tempDf = 1.1 % plusOne;
-            Assert.Equal(tempDf[0, 0], 1.1 % (double)plusOne[0, 0]);
-            tempDf = 1.1m % plusDecimal;
-            Assert.Equal(tempDf[0, 0], 1.1m % (decimal)plusDecimal[0, 0]);
-            Assert.True(typeof(decimal) == tempDf.Columns["Int"].DataType);
-
-            Assert.Equal((byte)0, df[0, 0]);
-        }
-
-        [Fact]
-        public void TestBinaryOperationsOnColumns()
-        {
-            Int32DataFrameColumn column = new Int32DataFrameColumn("Int", Enumerable.Range(0, 10));
-            Assert.ThrowsAny<ArgumentException>(() => column.Add(5.5, inPlace: true));
-            Assert.ThrowsAny<ArgumentException>(() => column.ReverseAdd(5.5, inPlace: true));
-            string str = "A String";
-            Assert.ThrowsAny<ArgumentException>(() => column.Add(str, inPlace: true));
-            Assert.ThrowsAny<ArgumentException>(() => column.ReverseAdd(str, inPlace: true));
-        }
-
-        [Fact]
         public void TestColumnReverseOrderState()
         {
             var column = new Int32DataFrameColumn("Int", Enumerable.Range(0, 10));
@@ -891,467 +217,6 @@ namespace Microsoft.Data.Analysis.Tests
 
             df["Int3"] = df.Columns["Int1"] * 2 + df.Columns["Int2"];
             Assert.Equal(16, df.Columns["Int3"][2]);
-        }
-
-        [Fact]
-        public void TestComputations()
-        {
-            DataFrame df = MakeDataFrameWithAllMutableColumnTypes(10);
-            df["Int"][0] = -10;
-            Assert.Equal(-10, df.Columns["Int"][0]);
-
-            DataFrameColumn absColumn = df.Columns["Int"].Abs();
-            Assert.Equal(10, absColumn[0]);
-            Assert.Equal(-10, df.Columns["Int"][0]);
-            df.Columns["Int"].Abs(true);
-            Assert.Equal(10, df.Columns["Int"][0]);
-
-            Assert.Throws<NotSupportedException>(() => df.Columns["Byte"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Byte"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Char"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Char"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["DateTime"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["DateTime"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Decimal"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Decimal"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Double"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Double"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Float"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Float"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Int"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Int"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Long"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Long"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Sbyte"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Sbyte"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Short"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Short"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Uint"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Uint"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Ulong"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Ulong"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Ushort"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Ushort"].Any());
-
-            bool any = df.Columns["Bool"].Any();
-            bool all = df.Columns["Bool"].All();
-            Assert.True(any);
-            Assert.False(all);
-
-            // Test the computation results
-            df.Columns["Double"][0] = 100.0;
-            DataFrameColumn doubleColumn = df.Columns["Double"].CumulativeMax();
-            for (int i = 0; i < doubleColumn.Length; i++)
-            {
-                if (i == 5)
-                    Assert.Null(doubleColumn[i]);
-                else
-                    Assert.Equal(100.0, (double)doubleColumn[i]);
-            }
-            Assert.Equal(1.0, df.Columns["Double"][1]);
-            df.Columns["Double"].CumulativeMax(true);
-            for (int i = 0; i < df.Columns["Double"].Length; i++)
-            {
-                if (i == 5)
-                    Assert.Null(df.Columns["Double"][i]);
-                else
-                    Assert.Equal(100.0, (double)df.Columns["Double"][i]);
-            }
-
-            df.Columns["Float"][0] = -10.0f;
-            DataFrameColumn floatColumn = df.Columns["Float"].CumulativeMin();
-            for (int i = 0; i < floatColumn.Length; i++)
-            {
-                if (i == 5)
-                    Assert.Null(floatColumn[i]);
-                else
-                    Assert.Equal(-10.0f, (float)floatColumn[i]);
-            }
-            Assert.Equal(9.0f, df.Columns["Float"][9]);
-            df.Columns["Float"].CumulativeMin(true);
-            for (int i = 0; i < df.Columns["Float"].Length; i++)
-            {
-                if (i == 5)
-                    Assert.Null(df.Columns["Float"][i]);
-                else
-                    Assert.Equal(-10.0f, (float)df.Columns["Float"][i]);
-            }
-
-            DataFrameColumn uintColumn = df.Columns["Uint"].CumulativeProduct();
-            Assert.Equal((uint)0, uintColumn[8]);
-            Assert.Equal((uint)8, df.Columns["Uint"][8]);
-            df.Columns["Uint"].CumulativeProduct(true);
-            Assert.Equal((uint)0, df.Columns["Uint"][9]);
-
-            DataFrameColumn ushortColumn = df.Columns["Ushort"].CumulativeSum();
-            Assert.Equal((ushort)40, ushortColumn[9]);
-            Assert.Equal((ushort)9, df.Columns["Ushort"][9]);
-            df.Columns["Ushort"].CumulativeSum(true);
-            Assert.Equal((ushort)40, df.Columns["Ushort"][9]);
-
-            Assert.Equal(100.0, df.Columns["Double"].Max());
-            Assert.Equal(-10.0f, df.Columns["Float"].Min());
-            Assert.Equal((uint)0, df.Columns["Uint"].Product());
-            Assert.Equal((ushort)130, df.Columns["Ushort"].Sum());
-
-            df.Columns["Double"][0] = 100.1;
-            Assert.Equal(100.1, df.Columns["Double"][0]);
-            DataFrameColumn roundColumn = df.Columns["Double"].Round();
-            Assert.Equal(100.0, roundColumn[0]);
-            Assert.Equal(100.1, df.Columns["Double"][0]);
-            df.Columns["Double"].Round(true);
-            Assert.Equal(100.0, df.Columns["Double"][0]);
-
-            // Test that none of the numeric column types throw
-            for (int i = 0; i < df.Columns.Count; i++)
-            {
-                DataFrameColumn column = df.Columns[i];
-                if (column.DataType == typeof(bool))
-                {
-                    Assert.Throws<NotSupportedException>(() => column.CumulativeMax());
-                    Assert.Throws<NotSupportedException>(() => column.CumulativeMin());
-                    Assert.Throws<NotSupportedException>(() => column.CumulativeProduct());
-                    Assert.Throws<NotSupportedException>(() => column.CumulativeSum());
-                    Assert.Throws<NotSupportedException>(() => column.Max());
-                    Assert.Throws<NotSupportedException>(() => column.Min());
-                    Assert.Throws<NotSupportedException>(() => column.Product());
-                    Assert.Throws<NotSupportedException>(() => column.Sum());
-                    continue;
-                }
-                else if (column.DataType == typeof(string))
-                {
-                    Assert.Throws<NotImplementedException>(() => column.CumulativeMax());
-                    Assert.Throws<NotImplementedException>(() => column.CumulativeMin());
-                    Assert.Throws<NotImplementedException>(() => column.CumulativeProduct());
-                    Assert.Throws<NotImplementedException>(() => column.CumulativeSum());
-                    Assert.Throws<NotImplementedException>(() => column.Max());
-                    Assert.Throws<NotImplementedException>(() => column.Min());
-                    Assert.Throws<NotImplementedException>(() => column.Product());
-                    Assert.Throws<NotImplementedException>(() => column.Sum());
-                    continue;
-                }
-                else if (column.DataType == typeof(DateTime))
-                {
-                    column.CumulativeMax();
-                    column.CumulativeMin();
-                    column.Max();
-                    column.Min();
-
-                    Assert.Throws<NotSupportedException>(() => column.CumulativeProduct());
-                    Assert.Throws<NotSupportedException>(() => column.CumulativeSum());
-                    Assert.Throws<NotSupportedException>(() => column.Product());
-                    Assert.Throws<NotSupportedException>(() => column.Sum());
-                    continue;
-                }
-
-                column.CumulativeMax();
-                column.CumulativeMin();
-                column.CumulativeProduct();
-                column.CumulativeSum();
-                column.Max();
-                column.Min();
-                column.Product();
-                column.Sum();
-            }
-        }
-
-        [Fact]
-        public void TestComputationsIncludingDateTime()
-        {
-            DataFrame df = MakeDataFrameWithNumericStringAndDateTimeColumns(10);
-            df["Int"][0] = -10;
-            Assert.Equal(-10, df.Columns["Int"][0]);
-
-            DataFrameColumn absColumn = df.Columns["Int"].Abs();
-            Assert.Equal(10, absColumn[0]);
-            Assert.Equal(-10, df.Columns["Int"][0]);
-            df.Columns["Int"].Abs(true);
-            Assert.Equal(10, df.Columns["Int"][0]);
-
-            Assert.Throws<NotSupportedException>(() => df.Columns["Byte"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Byte"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Char"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Char"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Decimal"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Decimal"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Double"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Double"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Float"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Float"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Int"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Int"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Long"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Long"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Sbyte"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Sbyte"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Short"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Short"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Uint"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Uint"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Ulong"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Ulong"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Ushort"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["Ushort"].Any());
-            Assert.Throws<NotSupportedException>(() => df.Columns["DateTime"].All());
-            Assert.Throws<NotSupportedException>(() => df.Columns["DateTime"].Any());
-
-            // Test the computation results
-            var maxDate = SampleDateTime.AddDays(100);
-            df.Columns["DateTime"][0] = maxDate;
-            DataFrameColumn dateTimeColumn = df.Columns["DateTime"].CumulativeMax();
-            for (int i = 0; i < dateTimeColumn.Length; i++)
-            {
-                if (i == 5)
-                    Assert.Null(dateTimeColumn[i]);
-                else
-                    Assert.Equal(maxDate, (DateTime)dateTimeColumn[i]);
-            }
-            Assert.Equal(maxDate, dateTimeColumn.Max());
-
-            df.Columns["Double"][0] = 100.0;
-            DataFrameColumn doubleColumn = df.Columns["Double"].CumulativeMax();
-            for (int i = 0; i < doubleColumn.Length; i++)
-            {
-                if (i == 5)
-                    Assert.Null(doubleColumn[i]);
-                else
-                    Assert.Equal(100.0, (double)doubleColumn[i]);
-            }
-            Assert.Equal(1.0, df.Columns["Double"][1]);
-            df.Columns["Double"].CumulativeMax(true);
-            for (int i = 0; i < df.Columns["Double"].Length; i++)
-            {
-                if (i == 5)
-                    Assert.Null(df.Columns["Double"][i]);
-                else
-                    Assert.Equal(100.0, (double)df.Columns["Double"][i]);
-            }
-
-            df.Columns["Float"][0] = -10.0f;
-            DataFrameColumn floatColumn = df.Columns["Float"].CumulativeMin();
-            for (int i = 0; i < floatColumn.Length; i++)
-            {
-                if (i == 5)
-                    Assert.Null(floatColumn[i]);
-                else
-                    Assert.Equal(-10.0f, (float)floatColumn[i]);
-            }
-            Assert.Equal(9.0f, df.Columns["Float"][9]);
-            df.Columns["Float"].CumulativeMin(true);
-            for (int i = 0; i < df.Columns["Float"].Length; i++)
-            {
-                if (i == 5)
-                    Assert.Null(df.Columns["Float"][i]);
-                else
-                    Assert.Equal(-10.0f, (float)df.Columns["Float"][i]);
-            }
-
-            DataFrameColumn uintColumn = df.Columns["Uint"].CumulativeProduct();
-            Assert.Equal((uint)0, uintColumn[8]);
-            Assert.Equal((uint)8, df.Columns["Uint"][8]);
-            df.Columns["Uint"].CumulativeProduct(true);
-            Assert.Equal((uint)0, df.Columns["Uint"][9]);
-
-            DataFrameColumn ushortColumn = df.Columns["Ushort"].CumulativeSum();
-            Assert.Equal((ushort)40, ushortColumn[9]);
-            Assert.Equal((ushort)9, df.Columns["Ushort"][9]);
-            df.Columns["Ushort"].CumulativeSum(true);
-            Assert.Equal((ushort)40, df.Columns["Ushort"][9]);
-
-            Assert.Equal(100.0, df.Columns["Double"].Max());
-            Assert.Equal(-10.0f, df.Columns["Float"].Min());
-            Assert.Equal((uint)0, df.Columns["Uint"].Product());
-            Assert.Equal((ushort)130, df.Columns["Ushort"].Sum());
-
-            df.Columns["Double"][0] = 100.1;
-            Assert.Equal(100.1, df.Columns["Double"][0]);
-            DataFrameColumn roundColumn = df.Columns["Double"].Round();
-            Assert.Equal(100.0, roundColumn[0]);
-            Assert.Equal(100.1, df.Columns["Double"][0]);
-            df.Columns["Double"].Round(true);
-            Assert.Equal(100.0, df.Columns["Double"][0]);
-
-            // Test that none of the numeric column types throw
-            for (int i = 0; i < df.Columns.Count; i++)
-            {
-                DataFrameColumn column = df.Columns[i];
-                if (column.DataType == typeof(bool))
-                {
-                    Assert.Throws<NotSupportedException>(() => column.CumulativeMax());
-                    Assert.Throws<NotSupportedException>(() => column.CumulativeMin());
-                    Assert.Throws<NotSupportedException>(() => column.CumulativeProduct());
-                    Assert.Throws<NotSupportedException>(() => column.CumulativeSum());
-                    Assert.Throws<NotSupportedException>(() => column.Max());
-                    Assert.Throws<NotSupportedException>(() => column.Min());
-                    Assert.Throws<NotSupportedException>(() => column.Product());
-                    Assert.Throws<NotSupportedException>(() => column.Sum());
-                    continue;
-                }
-                else if (column.DataType == typeof(string))
-                {
-                    Assert.Throws<NotImplementedException>(() => column.CumulativeMax());
-                    Assert.Throws<NotImplementedException>(() => column.CumulativeMin());
-                    Assert.Throws<NotImplementedException>(() => column.CumulativeProduct());
-                    Assert.Throws<NotImplementedException>(() => column.CumulativeSum());
-                    Assert.Throws<NotImplementedException>(() => column.Max());
-                    Assert.Throws<NotImplementedException>(() => column.Min());
-                    Assert.Throws<NotImplementedException>(() => column.Product());
-                    Assert.Throws<NotImplementedException>(() => column.Sum());
-                    continue;
-                }
-                else if (column.DataType == typeof(DateTime))
-                {
-                    Assert.Throws<NotSupportedException>(() => column.CumulativeProduct());
-                    Assert.Throws<NotSupportedException>(() => column.CumulativeSum());
-                    Assert.Throws<NotSupportedException>(() => column.Product());
-                    Assert.Throws<NotSupportedException>(() => column.Sum());
-                    continue;
-                }
-                column.CumulativeMax();
-                column.CumulativeMin();
-                column.CumulativeProduct();
-                column.CumulativeSum();
-                column.Max();
-                column.Min();
-                column.Product();
-                column.Sum();
-            }
-        }
-
-        [Fact]
-        public void TestIntComputations_MaxMin_WithNulls()
-        {
-            var column = new Int32DataFrameColumn("Int", new int?[]
-                {
-                    null,
-                    2,
-                    1,
-                    4,
-                    3,
-                    null
-                });
-
-            Assert.Equal(1, column.Min());
-            Assert.Equal(4, column.Max());
-        }
-
-        [Fact]
-        public void TestIntSum_OnColumnWithNullsOnly()
-        {
-            var column = new Int32DataFrameColumn("Int", new int?[] { null, null });
-            Assert.Null(column.Sum());
-        }
-
-        [Fact]
-        public void TestIntSum_OnEmptyColumn()
-        {
-            var column = new Int32DataFrameColumn("Int");
-            Assert.Null(column.Sum());
-        }
-
-        [Fact]
-        public void TestIntComputations_MaxMin_OnEmptyColumn()
-        {
-            var column = new Int32DataFrameColumn("Int");
-
-            Assert.Null(column.Min());
-            Assert.Null(column.Max());
-        }
-
-        [Fact]
-        public void TestDateTimeComputations_MaxMin_OnEmptyColumn()
-        {
-            var column = new DateTimeDataFrameColumn("DateTime");
-
-            Assert.Null(column.Min());
-            Assert.Null(column.Max());
-        }
-
-        [Fact]
-        public void TestDateTimeComputations_MaxMin_WithNulls()
-        {
-            var dateTimeColumn = new DateTimeDataFrameColumn("DateTime", new DateTime?[]
-                {
-                    null,
-                    new DateTime(2022, 1, 1),
-                    new DateTime(2020, 1, 1),
-                    new DateTime(2023, 1, 1),
-                    new DateTime(2021, 1, 1),
-                    null
-                });
-
-            Assert.Equal(new DateTime(2020, 1, 1), dateTimeColumn.Min());
-            Assert.Equal(new DateTime(2023, 1, 1), dateTimeColumn.Max());
-        }
-
-        [Theory]
-        [InlineData(5, 10)]
-        [InlineData(-15, 10)]
-        [InlineData(-5, 10)]
-        public void TestComputations_WithNegativeNumbers_MaxMin_Calculated(int startingFrom, int length)
-        {
-            // Arrange
-
-            IEnumerable<int> range = Enumerable.Range(startingFrom, length);
-
-            int max = range.Max();
-            int min = range.Min();
-
-            DataFrame df = MakeDataFrameWithNumericColumns(length, withNulls: false, startingFrom);
-
-            var byteColumn = (PrimitiveDataFrameColumn<byte>)df.Columns["Byte"];
-            var decimalColumn = (PrimitiveDataFrameColumn<decimal>)df.Columns["Decimal"];
-            var doubleColumn = (PrimitiveDataFrameColumn<double>)df.Columns["Double"];
-            var floatColumn = (PrimitiveDataFrameColumn<float>)df.Columns["Float"];
-            var intColumn = (PrimitiveDataFrameColumn<int>)df.Columns["Int"];
-            var longColumn = (PrimitiveDataFrameColumn<long>)df.Columns["Long"];
-            var sbyteColumn = (PrimitiveDataFrameColumn<sbyte>)df.Columns["Sbyte"];
-            var shortColumn = (PrimitiveDataFrameColumn<short>)df.Columns["Short"];
-            var uintColumn = (PrimitiveDataFrameColumn<uint>)df.Columns["Uint"];
-            var ulongColumn = (PrimitiveDataFrameColumn<ulong>)df.Columns["Ulong"];
-            var ushortColumn = (PrimitiveDataFrameColumn<ushort>)df.Columns["Ushort"];
-
-            // Act, Assert
-
-            // We need to iterate over all range with conversion to byte due to negative numbers issue
-            Assert.Equal((byte)byteColumn.Max(), range.Select(x => (byte)x).Max());
-
-            Assert.Equal((decimal)decimalColumn.Max(), (decimal)max);
-            Assert.Equal((double)doubleColumn.Max(), (double)max);
-            Assert.Equal((float)floatColumn.Max(), (float)max);
-            Assert.Equal((int)intColumn.Max(), (int)max);
-            Assert.Equal((long)longColumn.Max(), (long)max);
-            Assert.Equal((sbyte)sbyteColumn.Max(), (sbyte)max);
-            Assert.Equal((short)shortColumn.Max(), (short)max);
-
-            // We need to iterate over all range with conversion to uint due to negative numbers issue
-            Assert.Equal((uint)uintColumn.Max(), range.Select(x => (uint)x).Max());
-
-            // We need to iterate over all range with conversion to ulong due to negative numbers issue
-            Assert.Equal((ulong)ulongColumn.Max(), range.Select(x => (ulong)x).Max());
-
-            // We need to iterate over all range with conversion to ushort due to negative numbers issue
-            Assert.Equal((ushort)ushortColumn.Max(), range.Select(x => (ushort)x).Max());
-
-            // We need to iterate over all range with conversion to byte due to negative numbers issue
-            Assert.Equal((byte)byteColumn.Min(), range.Select(x => (byte)x).Min());
-
-            Assert.Equal((decimal)decimalColumn.Min(), (decimal)min);
-            Assert.Equal((double)doubleColumn.Min(), (double)min);
-            Assert.Equal((float)floatColumn.Min(), (float)min);
-            Assert.Equal((int)intColumn.Min(), (int)min);
-            Assert.Equal((long)longColumn.Min(), (long)min);
-            Assert.Equal((sbyte)sbyteColumn.Min(), (sbyte)min);
-            Assert.Equal((short)shortColumn.Min(), (short)min);
-
-            // We need to iterate over all range with conversion to uint due to negative numbers issue
-            Assert.Equal((uint)uintColumn.Min(), range.Select(x => (uint)x).Min());
-
-            // We need to iterate over all range with conversion to ulong due to negative numbers issue
-            Assert.Equal((ulong)ulongColumn.Min(), range.Select(x => (ulong)x).Min());
-
-            // We need to iterate over all range with conversion to ushort due to negative numbers issue
-            Assert.Equal((ushort)ushortColumn.Min(), range.Select(x => (ushort)x).Min());
         }
 
         [Fact]
@@ -1388,316 +253,6 @@ namespace Microsoft.Data.Analysis.Tests
             Assert.Null(sortedDf.Columns["Int"][0]);
             Assert.Equal(8, sortedDf.Columns["Int"][2]);
             Assert.Equal(9, sortedDf.Columns["Int"][1]);
-        }
-
-        [Fact]
-        public void TestSplitAndSort()
-        {
-            DataFrame df = MakeDataFrameWithAllMutableColumnTypes(20);
-            df.Columns["Int"][0] = 100000;
-            df.Columns["Int"][df.Rows.Count - 1] = -1;
-            df.Columns["Int"][5] = 200000;
-            DataFrame dfTest;
-            DataFrame dfTrain = SplitTrainTest(df, 0.8f, out dfTest);
-
-            // Sort by "Int" in ascending order
-            var sortedDf = dfTrain.OrderBy("Int");
-            Assert.Null(sortedDf.Columns["Int"][sortedDf.Rows.Count - 1]);
-            Assert.Equal(1, sortedDf.Columns["Int"][0]);
-            Assert.Equal(100000, sortedDf.Columns["Int"][sortedDf.Rows.Count - 3]);
-            Assert.Equal(200000, sortedDf.Columns["Int"][sortedDf.Rows.Count - 2]);
-        }
-
-        [Fact]
-        public void TestStringColumnSort()
-        {
-            // StringDataFrameColumn specific sort tests
-            StringDataFrameColumn strColumn = new StringDataFrameColumn("String", 0);
-            Assert.Equal(0, strColumn.NullCount);
-            for (int i = 0; i < 5; i++)
-            {
-                strColumn.Append(null);
-            }
-            Assert.Equal(5, strColumn.NullCount);
-            // Should handle all nulls
-            StringDataFrameColumn sortedStrColumn = strColumn.Sort() as StringDataFrameColumn;
-            Assert.Equal(5, sortedStrColumn.NullCount);
-            Assert.Null(sortedStrColumn[0]);
-
-            for (int i = 0; i < 5; i++)
-            {
-                strColumn.Append(i.ToString());
-            }
-            Assert.Equal(5, strColumn.NullCount);
-
-            // Ascending sort
-            sortedStrColumn = strColumn.Sort() as StringDataFrameColumn;
-            Assert.Equal("0", sortedStrColumn[0]);
-            Assert.Null(sortedStrColumn[9]);
-
-            // Descending sort
-            sortedStrColumn = strColumn.Sort(false) as StringDataFrameColumn;
-            Assert.Equal("4", sortedStrColumn[0]);
-            Assert.Null(sortedStrColumn[9]);
-        }
-
-        [Theory]
-        [InlineData(5)]
-        [InlineData(12)]
-        [InlineData(100)]
-        [InlineData(1000)]
-        public void TestPrimitiveColumnSort(int numberOfNulls)
-        {
-            // Primitive Column Sort
-            Int32DataFrameColumn intColumn = new Int32DataFrameColumn("Int", 0);
-            Assert.Equal(0, intColumn.NullCount);
-            intColumn.AppendMany(null, numberOfNulls);
-            Assert.Equal(numberOfNulls, intColumn.NullCount);
-
-            // Should handle all nulls
-            PrimitiveDataFrameColumn<int> sortedIntColumn = intColumn.Sort();
-            Assert.Equal(numberOfNulls, sortedIntColumn.NullCount);
-            Assert.Null(sortedIntColumn[0]);
-
-            for (int i = 0; i < 5; i++)
-            {
-                intColumn.Append(i);
-            }
-            Assert.Equal(numberOfNulls, intColumn.NullCount);
-
-            // Ascending sort
-            sortedIntColumn = intColumn.Sort();
-            Assert.Equal(0, sortedIntColumn[0]);
-            Assert.Null(sortedIntColumn[9]);
-
-            // Descending sort
-            sortedIntColumn = intColumn.Sort(ascending: false);
-            Assert.Equal(4, sortedIntColumn[0]);
-            Assert.Null(sortedIntColumn[9]);
-        }
-
-        [Fact]
-        public void TestSortWithDifferentNullCountsInColumns()
-        {
-            DataFrame dataFrame = MakeDataFrameWithAllMutableColumnTypes(10);
-            dataFrame["Int"][3] = null;
-            dataFrame["String"][3] = null;
-            DataFrame sorted = dataFrame.OrderBy("Int");
-            void Verify(DataFrame sortedDataFrame)
-            {
-                Assert.Equal(10, sortedDataFrame.Rows.Count);
-                DataFrameRow lastRow = sortedDataFrame.Rows[sortedDataFrame.Rows.Count - 1];
-                DataFrameRow penultimateRow = sortedDataFrame.Rows[sortedDataFrame.Rows.Count - 2];
-                foreach (object value in lastRow)
-                {
-                    Assert.Null(value);
-                }
-
-                for (int i = 0; i < sortedDataFrame.Columns.Count; i++)
-                {
-                    string columnName = sortedDataFrame.Columns[i].Name;
-                    if (columnName != "String" && columnName != "Int")
-                    {
-                        Assert.Equal(dataFrame[columnName][3], penultimateRow[i]);
-                    }
-                    else if (columnName == "String" || columnName == "Int")
-                    {
-                        Assert.Null(penultimateRow[i]);
-                    }
-                }
-            }
-
-            Verify(sorted);
-
-            sorted = dataFrame.OrderBy("String");
-            Verify(sorted);
-        }
-
-        private void VerifyJoin(DataFrame join, DataFrame left, DataFrame right, JoinAlgorithm joinAlgorithm)
-        {
-            Int64DataFrameColumn mapIndices = new Int64DataFrameColumn("map", join.Rows.Count);
-            for (long i = 0; i < join.Rows.Count; i++)
-            {
-                mapIndices[i] = i;
-            }
-            for (int i = 0; i < join.Columns.Count; i++)
-            {
-                DataFrameColumn joinColumn = join.Columns[i];
-                DataFrameColumn isEqual;
-
-                if (joinAlgorithm == JoinAlgorithm.Left)
-                {
-                    if (i < left.Columns.Count)
-                    {
-                        DataFrameColumn leftColumn = left.Columns[i];
-                        isEqual = joinColumn.ElementwiseEquals(leftColumn);
-                    }
-                    else
-                    {
-                        int columnIndex = i - left.Columns.Count;
-                        DataFrameColumn rightColumn = right.Columns[columnIndex];
-                        DataFrameColumn compareColumn = rightColumn.Length <= join.Rows.Count ? rightColumn.Clone(numberOfNullsToAppend: join.Rows.Count - rightColumn.Length) : rightColumn.Clone(mapIndices);
-                        isEqual = joinColumn.ElementwiseEquals(compareColumn);
-                    }
-                }
-                else if (joinAlgorithm == JoinAlgorithm.Right)
-                {
-                    if (i < left.Columns.Count)
-                    {
-                        DataFrameColumn leftColumn = left.Columns[i];
-                        DataFrameColumn compareColumn = leftColumn.Length <= join.Rows.Count ? leftColumn.Clone(numberOfNullsToAppend: join.Rows.Count - leftColumn.Length) : leftColumn.Clone(mapIndices);
-                        isEqual = joinColumn.ElementwiseEquals(compareColumn);
-                    }
-                    else
-                    {
-                        int columnIndex = i - left.Columns.Count;
-                        DataFrameColumn rightColumn = right.Columns[columnIndex];
-                        isEqual = joinColumn.ElementwiseEquals(rightColumn);
-                    }
-                }
-                else if (joinAlgorithm == JoinAlgorithm.Inner)
-                {
-                    if (i < left.Columns.Count)
-                    {
-                        DataFrameColumn leftColumn = left.Columns[i];
-                        isEqual = joinColumn.ElementwiseEquals(leftColumn.Clone(mapIndices));
-                    }
-                    else
-                    {
-                        int columnIndex = i - left.Columns.Count;
-                        DataFrameColumn rightColumn = right.Columns[columnIndex];
-                        isEqual = joinColumn.ElementwiseEquals(rightColumn.Clone(mapIndices));
-                    }
-                }
-                else
-                {
-                    if (i < left.Columns.Count)
-                    {
-                        DataFrameColumn leftColumn = left.Columns[i];
-                        isEqual = joinColumn.ElementwiseEquals(leftColumn.Clone(numberOfNullsToAppend: join.Rows.Count - leftColumn.Length));
-                    }
-                    else
-                    {
-                        int columnIndex = i - left.Columns.Count;
-                        DataFrameColumn rightColumn = right.Columns[columnIndex];
-                        isEqual = joinColumn.ElementwiseEquals(rightColumn.Clone(numberOfNullsToAppend: join.Rows.Count - rightColumn.Length));
-                    }
-                }
-                for (int j = 0; j < join.Rows.Count; j++)
-                {
-                    Assert.Equal(true, isEqual[j]);
-                }
-            }
-        }
-
-        private void VerifyMerge(DataFrame merge, DataFrame left, DataFrame right, JoinAlgorithm joinAlgorithm)
-        {
-            if (joinAlgorithm == JoinAlgorithm.Left || joinAlgorithm == JoinAlgorithm.Inner)
-            {
-                HashSet<int> intersection = new HashSet<int>();
-                for (int i = 0; i < merge.Columns["Int_left"].Length; i++)
-                {
-                    if (merge.Columns["Int_left"][i] == null)
-                        continue;
-                    intersection.Add((int)merge.Columns["Int_left"][i]);
-                }
-                for (int i = 0; i < left.Columns["Int"].Length; i++)
-                {
-                    if (left.Columns["Int"][i] != null && intersection.Contains((int)left.Columns["Int"][i]))
-                        intersection.Remove((int)left.Columns["Int"][i]);
-                }
-                Assert.Empty(intersection);
-            }
-            else if (joinAlgorithm == JoinAlgorithm.Right)
-            {
-                HashSet<int> intersection = new HashSet<int>();
-                for (int i = 0; i < merge.Columns["Int_right"].Length; i++)
-                {
-                    if (merge.Columns["Int_right"][i] == null)
-                        continue;
-                    intersection.Add((int)merge.Columns["Int_right"][i]);
-                }
-                for (int i = 0; i < right.Columns["Int"].Length; i++)
-                {
-                    if (right.Columns["Int"][i] != null && intersection.Contains((int)right.Columns["Int"][i]))
-                        intersection.Remove((int)right.Columns["Int"][i]);
-                }
-                Assert.Empty(intersection);
-            }
-            else if (joinAlgorithm == JoinAlgorithm.FullOuter)
-            {
-                VerifyMerge(merge, left, right, JoinAlgorithm.Left);
-                VerifyMerge(merge, left, right, JoinAlgorithm.Right);
-            }
-        }
-
-        [Fact]
-        public void TestJoin()
-        {
-            DataFrame left = MakeDataFrameWithAllMutableColumnTypes(10);
-            DataFrame right = MakeDataFrameWithAllMutableColumnTypes(5);
-
-            // Tests with right.Rows.Count < left.Rows.Count
-            // Left join
-            DataFrame join = left.Join(right);
-            Assert.Equal(join.Rows.Count, left.Rows.Count);
-            Assert.Equal(join.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Null(join.Columns["Int_right"][6]);
-            VerifyJoin(join, left, right, JoinAlgorithm.Left);
-
-            // Right join
-            join = left.Join(right, joinAlgorithm: JoinAlgorithm.Right);
-            Assert.Equal(join.Rows.Count, right.Rows.Count);
-            Assert.Equal(join.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Equal(join.Columns["Int_right"][3], right.Columns["Int"][3]);
-            Assert.Null(join.Columns["Int_right"][2]);
-            VerifyJoin(join, left, right, JoinAlgorithm.Right);
-
-            // Outer join
-            join = left.Join(right, joinAlgorithm: JoinAlgorithm.FullOuter);
-            Assert.Equal(join.Rows.Count, left.Rows.Count);
-            Assert.Equal(join.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Null(join.Columns["Int_right"][6]);
-            VerifyJoin(join, left, right, JoinAlgorithm.FullOuter);
-
-            // Inner join
-            join = left.Join(right, joinAlgorithm: JoinAlgorithm.Inner);
-            Assert.Equal(join.Rows.Count, right.Rows.Count);
-            Assert.Equal(join.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Equal(join.Columns["Int_right"][3], right.Columns["Int"][3]);
-            Assert.Null(join.Columns["Int_right"][2]);
-            VerifyJoin(join, left, right, JoinAlgorithm.Inner);
-
-            // Tests with right.Rows.Count > left.Rows.Count
-            // Left join
-            right = MakeDataFrameWithAllMutableColumnTypes(15);
-            join = left.Join(right);
-            Assert.Equal(join.Rows.Count, left.Rows.Count);
-            Assert.Equal(join.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Equal(join.Columns["Int_right"][6], right.Columns["Int"][6]);
-            VerifyJoin(join, left, right, JoinAlgorithm.Left);
-
-            // Right join
-            join = left.Join(right, joinAlgorithm: JoinAlgorithm.Right);
-            Assert.Equal(join.Rows.Count, right.Rows.Count);
-            Assert.Equal(join.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Equal(join.Columns["Int_right"][2], right.Columns["Int"][2]);
-            Assert.Null(join.Columns["Int_left"][12]);
-            VerifyJoin(join, left, right, JoinAlgorithm.Right);
-
-            // Outer join
-            join = left.Join(right, joinAlgorithm: JoinAlgorithm.FullOuter);
-            Assert.Equal(join.Rows.Count, right.Rows.Count);
-            Assert.Equal(join.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Null(join.Columns["Int_left"][12]);
-            VerifyJoin(join, left, right, JoinAlgorithm.FullOuter);
-
-            // Inner join
-            join = left.Join(right, joinAlgorithm: JoinAlgorithm.Inner);
-            Assert.Equal(join.Rows.Count, left.Rows.Count);
-            Assert.Equal(join.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Equal(join.Columns["Int_right"][2], right.Columns["Int"][2]);
-            VerifyJoin(join, left, right, JoinAlgorithm.Inner);
         }
 
         [Fact]
@@ -2003,18 +558,6 @@ namespace Microsoft.Data.Analysis.Tests
         }
 
         [Fact]
-        public void TestColumnFilter()
-        {
-            DataFrame df = MakeDataFrameWithNumericColumns(10);
-            DataFrameColumn filtered = df.Columns["Int"].Filter(3, 7);
-            Assert.Equal(4, filtered.Length);
-            Assert.Equal(3, filtered[0]);
-            Assert.Equal(4, filtered[1]);
-            Assert.Equal(6, filtered[2]);
-            Assert.Equal(7, filtered[3]);
-        }
-
-        [Fact]
         public void TestDataFrameClamp()
         {
             DataFrame df = MakeDataFrameWithAllColumnTypes(10);
@@ -2070,41 +613,6 @@ namespace Microsoft.Data.Analysis.Tests
             df.Clamp(3, 7, true);
             VerifyDataFrameClamp(df);
 
-        }
-
-        [Fact]
-        public void TestDataFrameFilter()
-        {
-            DataFrame df = MakeDataFrameWithAllMutableColumnTypes(10);
-            DataFrame boolColumnFiltered = df[df.Columns["Bool"].ElementwiseEquals(true)];
-            List<int> verify = new List<int> { 0, 2, 4, 6, 8 };
-            Assert.Equal(5, boolColumnFiltered.Rows.Count);
-            for (int i = 0; i < boolColumnFiltered.Columns.Count; i++)
-            {
-                DataFrameColumn column = boolColumnFiltered.Columns[i];
-                if (column.Name == "Char" || column.Name == "Bool" || column.Name == "String" || column.Name == "DateTime")
-                    continue;
-                for (int j = 0; j < column.Length; j++)
-                {
-                    Assert.Equal(verify[j].ToString(), column[j].ToString());
-                }
-            }
-            DataFrame intEnumerableFiltered = df[Enumerable.Range(0, 10)];
-            DataFrame boolEnumerableFiltered = df[Enumerable.Range(0, 10).Select(x => true)];
-            DataFrame longEnumerableFiltered = df[Enumerable.Range(0, 10).Select(x => (long)x)];
-            Assert.Equal(intEnumerableFiltered.Columns.Count, df.Columns.Count);
-            Assert.Equal(boolEnumerableFiltered.Columns.Count, df.Columns.Count);
-            Assert.Equal(longEnumerableFiltered.Columns.Count, df.Columns.Count);
-            for (int i = 0; i < intEnumerableFiltered.Columns.Count; i++)
-            {
-                DataFrameColumn intFilteredColumn = intEnumerableFiltered.Columns[i];
-                DataFrameColumn dfColumn = df.Columns[i];
-                DataFrameColumn boolFilteredColumn = boolEnumerableFiltered.Columns[i];
-                DataFrameColumn longFilteredColumn = longEnumerableFiltered.Columns[i];
-                Assert.True(intFilteredColumn.ElementwiseEquals(dfColumn).All());
-                Assert.True(boolFilteredColumn.ElementwiseEquals(dfColumn).All());
-                Assert.True(longFilteredColumn.ElementwiseEquals(dfColumn).All());
-            }
         }
 
         [Fact]
@@ -2170,757 +678,6 @@ namespace Microsoft.Data.Analysis.Tests
 
             // should throw exception as sample size is greater than dataframe rows
             Assert.Throws<ArgumentException>(() => df.Sample(13));
-        }
-
-        [Theory]
-        [InlineData(1, 2)]
-        [InlineData(2, 1)]
-        public void TestDataCorrectnessForInnerMerge(int leftCount, int rightCount)
-        {
-            DataFrame left = MakeDataFrameWithNumericColumns(leftCount, false);
-            DataFrameColumn leftStringColumn = new StringDataFrameColumn("String", Enumerable.Range(0, leftCount).Select(x => "Left"));
-            left.Columns.Insert(left.Columns.Count, leftStringColumn);
-
-            DataFrame right = MakeDataFrameWithNumericColumns(rightCount, false);
-            DataFrameColumn rightStringColumn = new StringDataFrameColumn("String", Enumerable.Range(0, rightCount).Select(x => "Right"));
-            right.Columns.Insert(right.Columns.Count, rightStringColumn);
-
-            DataFrame merge = left.Merge<int>(right, "Int", "Int", joinAlgorithm: JoinAlgorithm.Inner);
-
-            Assert.Equal("Left", (string)merge.Columns["String_left"][0]);
-            Assert.Equal("Right", (string)merge.Columns["String_right"][0]);
-        }
-
-        [Fact]
-        public void TestMerge()
-        {
-            DataFrame left = MakeDataFrameWithAllMutableColumnTypes(10);
-            DataFrame right = MakeDataFrameWithAllMutableColumnTypes(5);
-
-            // Tests with right.Rows.Count < left.Rows.Count
-            // Left merge
-            DataFrame merge = left.Merge<int>(right, "Int", "Int");
-            Assert.Equal(10, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Null(merge.Columns["Int_right"][6]);
-            Assert.Null(merge.Columns["Int_left"][5]);
-            VerifyMerge(merge, left, right, JoinAlgorithm.Left);
-
-            // Right merge
-            merge = left.Merge<int>(right, "Int", "Int", joinAlgorithm: JoinAlgorithm.Right);
-            Assert.Equal(5, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Equal(merge.Columns["Int_right"][3], right.Columns["Int"][3]);
-            Assert.Null(merge.Columns["Int_right"][2]);
-            VerifyMerge(merge, left, right, JoinAlgorithm.Right);
-
-            // Outer merge
-            merge = left.Merge<int>(right, "Int", "Int", joinAlgorithm: JoinAlgorithm.FullOuter);
-            Assert.Equal(merge.Rows.Count, left.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Null(merge.Columns["Int_right"][6]);
-            VerifyMerge(merge, left, right, JoinAlgorithm.FullOuter);
-
-            // Inner merge
-            merge = left.Merge<int>(right, "Int", "Int", joinAlgorithm: JoinAlgorithm.Inner);
-            Assert.Equal(merge.Rows.Count, right.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Equal(merge.Columns["Int_right"][2], right.Columns["Int"][3]);
-            Assert.Null(merge.Columns["Int_right"][4]);
-            VerifyMerge(merge, left, right, JoinAlgorithm.Inner);
-
-            // Tests with right.Rows.Count > left.Rows.Count
-            // Left merge
-            right = MakeDataFrameWithAllMutableColumnTypes(15);
-            merge = left.Merge<int>(right, "Int", "Int");
-            Assert.Equal(merge.Rows.Count, left.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Equal(merge.Columns["Int_right"][6], right.Columns["Int"][6]);
-            VerifyMerge(merge, left, right, JoinAlgorithm.Left);
-
-            // Right merge
-            merge = left.Merge<int>(right, "Int", "Int", joinAlgorithm: JoinAlgorithm.Right);
-            Assert.Equal(merge.Rows.Count, right.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Equal(merge.Columns["Int_right"][2], right.Columns["Int"][2]);
-            Assert.Null(merge.Columns["Int_left"][12]);
-            VerifyMerge(merge, left, right, JoinAlgorithm.Right);
-
-            // Outer merge
-            merge = left.Merge<int>(right, "Int", "Int", joinAlgorithm: JoinAlgorithm.FullOuter);
-            Assert.Equal(16, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Null(merge.Columns["Int_left"][12]);
-            Assert.Null(merge.Columns["Int_left"][15]);
-            VerifyMerge(merge, left, right, JoinAlgorithm.FullOuter);
-
-            // Inner merge
-            merge = left.Merge<int>(right, "Int", "Int", joinAlgorithm: JoinAlgorithm.Inner);
-            Assert.Equal(9, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-            Assert.Equal(merge.Columns["Int_right"][2], right.Columns["Int"][2]);
-            VerifyMerge(merge, left, right, JoinAlgorithm.Inner);
-        }
-
-        private void MatchRowsOnMergedDataFrame(DataFrame merge, DataFrame left, DataFrame right, long mergeRow, long? leftRow, long? rightRow)
-        {
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-            DataFrameRow dataFrameMergeRow = merge.Rows[mergeRow];
-            int columnIndex = 0;
-            foreach (object value in dataFrameMergeRow)
-            {
-                object compare = null;
-                if (columnIndex < left.Columns.Count)
-                {
-                    if (leftRow != null)
-                    {
-                        compare = left.Rows[leftRow.Value][columnIndex];
-                    }
-                }
-                else
-                {
-                    int rightColumnIndex = columnIndex - left.Columns.Count;
-                    if (rightRow != null)
-                    {
-                        compare = right.Rows[rightRow.Value][rightColumnIndex];
-                    }
-                }
-                Assert.Equal(value, compare);
-                columnIndex++;
-            }
-        }
-
-        [Theory]
-        [InlineData(10, 5, JoinAlgorithm.Left)]
-        [InlineData(5, 10, JoinAlgorithm.Right)]
-        public void TestMergeEdgeCases_LeftOrRight(int leftLength, int rightLength, JoinAlgorithm joinAlgorithm)
-        {
-            DataFrame left = MakeDataFrameWithAllMutableColumnTypes(leftLength);
-            if (leftLength > 5)
-            {
-                left["Int"][8] = null;
-            }
-            DataFrame right = MakeDataFrameWithAllMutableColumnTypes(rightLength);
-            if (rightLength > 5)
-            {
-                right["Int"][8] = null;
-            }
-
-            DataFrame merge = left.Merge<int>(right, "Int", "Int", joinAlgorithm: joinAlgorithm);
-            Assert.Equal(10, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-            int[] matchedFullRows = new int[] { 0, 1, 3, 4 };
-            for (long i = 0; i < matchedFullRows.Length; i++)
-            {
-                int rowIndex = matchedFullRows[i];
-                MatchRowsOnMergedDataFrame(merge, left, right, rowIndex, rowIndex, rowIndex);
-            }
-
-            int[] matchedLeftOrRightRowsNullOtherRows = new int[] { 2, 5, 6, 7, 8, 9 };
-            for (long i = 0; i < matchedLeftOrRightRowsNullOtherRows.Length; i++)
-            {
-                int rowIndex = matchedLeftOrRightRowsNullOtherRows[i];
-                if (leftLength > 5)
-                {
-                    MatchRowsOnMergedDataFrame(merge, left, right, rowIndex, rowIndex, null);
-                }
-                else
-                {
-                    MatchRowsOnMergedDataFrame(merge, left, right, rowIndex, null, rowIndex);
-                }
-            }
-        }
-
-        [Fact]
-        public void TestMergeEdgeCases_Inner()
-        {
-            DataFrame left = MakeDataFrameWithAllMutableColumnTypes(5);
-            DataFrame right = MakeDataFrameWithAllMutableColumnTypes(10);
-            left["Int"][3] = null;
-            right["Int"][6] = null;
-            // Creates this case:
-            /*
-             * Left:    Right:
-             * 0        0
-             * 1        1
-             * null(2)  2
-             * null(3)  3
-             * 4        4
-             *          null(5)
-             *          null(6)
-             *          7
-             *          8
-             *          9
-             */
-            /*
-             * Merge will result in a DataFrame like:
-             * Int_Left Int_Right
-             * 0        0
-             * 1        1
-             * 4        4
-             * null(2)  null(5)
-             * null(3)  null(5)
-             * null(2)  null(6)
-             * null(3)  null(6)
-             */
-
-            DataFrame merge = left.Merge<int>(right, "Int", "Int", joinAlgorithm: JoinAlgorithm.Inner);
-            Assert.Equal(7, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-
-            int[] mergeRows = new int[] { 0, 1, 2, 3, 4, 5, 6 };
-            int[] leftRows = new int[] { 0, 1, 4, 2, 3, 2, 3 };
-            int[] rightRows = new int[] { 0, 1, 4, 5, 5, 6, 6 };
-            for (long i = 0; i < mergeRows.Length; i++)
-            {
-                int rowIndex = mergeRows[i];
-                int leftRowIndex = leftRows[i];
-                int rightRowIndex = rightRows[i];
-                MatchRowsOnMergedDataFrame(merge, left, right, rowIndex, leftRowIndex, rightRowIndex);
-            }
-        }
-
-        [Fact]
-        public void TestMergeEdgeCases_Outer()
-        {
-            DataFrame left = MakeDataFrameWithAllMutableColumnTypes(5);
-            left["Int"][3] = null;
-            DataFrame right = MakeDataFrameWithAllMutableColumnTypes(5);
-            right["Int"][1] = 5;
-            right["Int"][3] = null;
-            right["Int"][4] = 6;
-
-            // Creates this case:
-            /*
-             * Left:    Right:    RowIndex:
-             * 0        0         0
-             * 1        5         1
-             * null     null      2
-             * null(3)  null(3)   3
-             * 4        6         4
-             */
-
-            /*
-             * Merge will result in a DataFrame like:
-             * Int_left:    Int_right:        Merged:    Index:
-             * 0            0                 0 - 0      0
-             * 1            null              1 - N      1
-             * null         null              2 - 2      2
-             * null         null(3)           2 - 3      3
-             * null(3)      null              3 - 2      4
-             * null(3)      null(3)           3 - 3      5
-             * 4            null              4 - N      6
-             * null         5                 N - 1      7
-             * null         6                 N - 4      8
-             */
-
-            DataFrame merge = left.Merge<int>(right, "Int", "Int", joinAlgorithm: JoinAlgorithm.FullOuter);
-            Assert.Equal(9, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-
-            int[] mergeRows = new int[] { 0, 2, 3, 4, 5 };
-            int[] leftRows = new int[] { 0, 2, 2, 3, 3 };
-            int[] rightRows = new int[] { 0, 2, 3, 2, 3 };
-            for (long i = 0; i < mergeRows.Length; i++)
-            {
-                int rowIndex = mergeRows[i];
-                int leftRowIndex = leftRows[i];
-                int rightRowIndex = rightRows[i];
-                MatchRowsOnMergedDataFrame(merge, left, right, rowIndex, leftRowIndex, rightRowIndex);
-            }
-
-            mergeRows = new int[] { 1, 6 };
-            leftRows = new int[] { 1, 4 };
-            for (long i = 0; i < mergeRows.Length; i++)
-            {
-                int rowIndex = mergeRows[i];
-                int leftRowIndex = leftRows[i];
-                MatchRowsOnMergedDataFrame(merge, left, right, rowIndex, leftRowIndex, null);
-            }
-
-            mergeRows = new int[] { 7, 8 };
-            rightRows = new int[] { 1, 4 };
-            for (long i = 0; i < mergeRows.Length; i++)
-            {
-                int rowIndex = mergeRows[i];
-                int rightRowIndex = rightRows[i];
-                MatchRowsOnMergedDataFrame(merge, left, right, rowIndex, null, rightRowIndex);
-            }
-        }
-
-        [Fact]
-        public void TestMerge_ByTwoColumns_Complex_LeftJoin()
-        {
-            //Test left merge by to int type columns
-
-            //Arrange
-            var left = new DataFrame();
-            left.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2, 3, 4, 5 }));
-            left.Columns.Add(new Int32DataFrameColumn("G1", new[] { 0, 1, 1, 2, 2, 3 }));
-            left.Columns.Add(new Int32DataFrameColumn("G2", new[] { 3, 1, 2, 1, 2, 1 }));
-
-            var right = new DataFrame();
-            right.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2, 3 }));
-            right.Columns.Add(new Int32DataFrameColumn("G1", new[] { 1, 1, 1, 2 }));
-            right.Columns.Add(new Int32DataFrameColumn("G2", new[] { 1, 2, 1, 1 }));
-
-            // Creates this case:
-            /*  -------------------------
-             *     Left     |     Right
-             *   I  G1 G2   |   I  G1 G2
-             *  -------------------------
-             *   0  0  3    |   0  1  1
-             *   1  1  1    |   1  1  2
-             *   2  1  2    |   2  1  1
-             *   3  2  1    |   3  2  1
-             *   4  2  2
-             *   5  3  1
-             */
-
-            /*
-             * Merge will result in a DataFrame like:
-             *   IL G1 G2     IR              Merged:
-             *  -------------------------
-             *   0  0  3                      0 - N
-             *   1  1  1       0  1  1        1 - 0
-             *   1  1  1       2  1  1        1 - 2
-             *   2  1  2       1  1  2        2 - 1
-             *   3  2  1       3  2  1        3 - 3
-             *   4  2  2                      4 - N
-             *   5  3  1                      5 - N
-             */
-
-            //Act
-            var merge = left.Merge(right, new[] { "G1", "G2" }, new[] { "G1", "G2" });
-
-            //Assert
-            var expectedMerged = new (int? Left, int? Right)[] {
-                (0, null),
-                (1, 0),
-                (1, 2),
-                (2, 1),
-                (3, 3),
-                (4, null),
-                (5, null)
-            };
-
-            Assert.Equal(expectedMerged.Length, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-
-            for (long i = 0; i < expectedMerged.Length; i++)
-            {
-                MatchRowsOnMergedDataFrame(merge, left, right, i, expectedMerged[i].Left, expectedMerged[i].Right);
-            }
-
-        }
-
-        [Fact]
-        public void TestMerge_ByTwoColumns_Simple_ManyToMany_LeftJoin()
-        {
-            //Test left merge by to int type columns
-
-            //Arrange
-            var left = new DataFrame();
-            left.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2 }));
-            left.Columns.Add(new Int32DataFrameColumn("G1", new[] { 1, 1, 3 }));
-            left.Columns.Add(new Int32DataFrameColumn("G2", new[] { 1, 1, 3 }));
-
-            var right = new DataFrame();
-            right.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2 }));
-            right.Columns.Add(new Int32DataFrameColumn("G1", new[] { 1, 1, 0 }));
-            right.Columns.Add(new Int32DataFrameColumn("G2", new[] { 1, 1, 0 }));
-
-            // Creates this case:
-            /*  ---------------------------
-             *     Left    |    Right
-             *   I  G1 G2  |   I  G1 G2
-             *  ---------------------------
-             *   0  1  1   |   0  1  1
-             *   1  1  1   |   1  1  1
-             *   2  3  3   |   2  0  0
-             */
-
-            /*
-             * Merge will result in a DataFrame like:
-             *   IL G1 G2     IR           Merged:
-             *  -------------------------
-             *   0  1  1      0  1  1       0 - 0
-             *   0  1  1      1  1  1       0 - 1
-             *   1  1  1      0  1  1       1 - 0
-             *   1  1  1      1  1  1       1 - 1
-             *   2  3  3                    2 - N
-             */
-
-            //Act
-            var merge = left.Merge(right, new[] { "G1", "G2" }, new[] { "G1", "G2" });
-
-            //Assert
-            var expectedMerged = new (int? Left, int? Right)[] {
-                (0, 0),
-                (0, 1),
-                (1, 0),
-                (1, 1),
-                (2, null)
-            };
-
-            Assert.Equal(expectedMerged.Length, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-
-            for (long i = 0; i < expectedMerged.Length; i++)
-            {
-                MatchRowsOnMergedDataFrame(merge, left, right, i, expectedMerged[i].Left, expectedMerged[i].Right);
-            }
-        }
-
-        [Fact]
-        public void TestMerge_ByTwoColumns_Simple_ManyToMany_RightJoin()
-        {
-            //Test left merge by to int type columns
-
-            //Arrange
-            var left = new DataFrame();
-            left.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2 }));
-            left.Columns.Add(new Int32DataFrameColumn("G1", new[] { 1, 1, 3 }));
-            left.Columns.Add(new Int32DataFrameColumn("G2", new[] { 1, 1, 3 }));
-
-            var right = new DataFrame();
-            right.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2 }));
-            right.Columns.Add(new Int32DataFrameColumn("G1", new[] { 1, 1, 0 }));
-            right.Columns.Add(new Int32DataFrameColumn("G2", new[] { 1, 1, 0 }));
-
-            // Creates this case:
-            /*  ---------------------------
-             *     Left    |    Right
-             *   I  G1 G2  |   I  G1 G2
-             *  ---------------------------
-             *   0  1  1   |   0  1  1
-             *   1  1  1   |   1  1  1
-             *   2  3  3   |   2  0  0
-             */
-
-            /*
-             * Merge will result in a DataFrame like:
-             *   IL G1 G2     IR           Merged:
-             *  -------------------------
-             *   0  1  1      0  1  1       0 - 0
-             *   1  1  1      0  1  1       1 - 0
-             *   0  1  1      1  1  1       0 - 1
-             *   1  1  1      1  1  1       1 - 1
-             *                2  0  0       N - 2
-             */
-
-            //Act
-            var merge = left.Merge(right, new[] { "G1", "G2" }, new[] { "G1", "G2" }, joinAlgorithm: JoinAlgorithm.Right);
-
-            //Assert
-            var expectedMerged = new (int? Left, int? Right)[] {
-                (0, 0),
-                (1, 0),
-                (0, 1),
-                (1, 1),
-                (null, 2)
-            };
-
-            Assert.Equal(expectedMerged.Length, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-
-            for (long i = 0; i < expectedMerged.Length; i++)
-            {
-                MatchRowsOnMergedDataFrame(merge, left, right, i, expectedMerged[i].Left, expectedMerged[i].Right);
-            }
-        }
-
-        [Fact]
-        public void TestMerge_ByTwoColumns_Simple_ManyToMany_InnerJoin()
-        {
-            //Test left merge by to int type columns
-
-            //Arrange
-            var left = new DataFrame();
-            left.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2 }));
-            left.Columns.Add(new Int32DataFrameColumn("G1", new[] { 1, 1, 3 }));
-            left.Columns.Add(new Int32DataFrameColumn("G2", new[] { 1, 1, 3 }));
-
-            var right = new DataFrame();
-            right.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2 }));
-            right.Columns.Add(new Int32DataFrameColumn("G1", new[] { 1, 1, 0 }));
-            right.Columns.Add(new Int32DataFrameColumn("G2", new[] { 1, 1, 0 }));
-
-            // Creates this case:
-            /*  ---------------------------
-             *     Left    |    Right
-             *   I  G1 G2  |   I  G1 G2
-             *  ---------------------------
-             *   0  1  1   |   0  1  1
-             *   1  1  1   |   1  1  1
-             *   2  3  3   |   2  0  0
-             */
-
-            /*
-             * Merge will result in a DataFrame like:
-             *   IL G1 G2     IR           Merged:
-             *  -------------------------
-             *   0  1  1      0  1  1       0 - 0
-             *   1  1  1      0  1  1       1 - 0
-             *   0  1  1      1  1  1       0 - 1
-             *   1  1  1      1  1  1       1 - 1
-             */
-
-            //Act
-            var merge = left.Merge(right, new[] { "G1", "G2" }, new[] { "G1", "G2" }, joinAlgorithm: JoinAlgorithm.Inner);
-
-            //Assert
-            var expectedMerged = new (int? Left, int? Right)[] {
-                (0, 0),
-                (1, 0),
-                (0, 1),
-                (1, 1)
-            };
-
-            Assert.Equal(expectedMerged.Length, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-
-            for (long i = 0; i < expectedMerged.Length; i++)
-            {
-                MatchRowsOnMergedDataFrame(merge, left, right, i, expectedMerged[i].Left, expectedMerged[i].Right);
-            }
-        }
-
-        [Fact]
-        public void TestMerge_ByTwoColumns_Simple_ManyToMany_OuterJoin()
-        {
-            //Test left merge by to int type columns
-
-            //Arrange
-            var left = new DataFrame();
-            left.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2 }));
-            left.Columns.Add(new Int32DataFrameColumn("G1", new[] { 1, 1, 3 }));
-            left.Columns.Add(new Int32DataFrameColumn("G2", new[] { 1, 1, 3 }));
-
-            var right = new DataFrame();
-            right.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2 }));
-            right.Columns.Add(new Int32DataFrameColumn("G1", new[] { 1, 1, 0 }));
-            right.Columns.Add(new Int32DataFrameColumn("G2", new[] { 1, 1, 0 }));
-
-            // Creates this case:
-            /*  ---------------------------
-             *     Left    |    Right
-             *   I  G1 G2  |   I  G1 G2
-             *  ---------------------------
-             *   0  1  1   |   0  1  1
-             *   1  1  1   |   1  1  1
-             *   2  3  3   |   2  0  0
-             */
-
-            /*
-             * Merge will result in a DataFrame like:
-             *   IL G1 G2     IR           Merged:
-             *  -------------------------
-             *   0  1  1      0  1  1       0 - 0
-             *   0  1  1      1  1  1       0 - 1
-             *   1  1  1      0  1  1       1 - 0
-             *   1  1  1      1  1  1       1 - 1
-             *   2  3  3                    2 - N
-             *                2  0  0       N - 2
-             */
-
-            //Act
-            var merge = left.Merge(right, new[] { "G1", "G2" }, new[] { "G1", "G2" }, joinAlgorithm: JoinAlgorithm.FullOuter);
-
-            //Assert
-            var expectedMerged = new (int? Left, int? Right)[] {
-                (0, 0),
-                (0, 1),
-                (1, 0),
-                (1, 1),
-                (2, null),
-                (null, 2)
-            };
-
-            Assert.Equal(expectedMerged.Length, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-
-            for (long i = 0; i < expectedMerged.Length; i++)
-            {
-                MatchRowsOnMergedDataFrame(merge, left, right, i, expectedMerged[i].Left, expectedMerged[i].Right);
-            }
-        }
-
-        [Fact]
-        public void TestMerge_ByThreeColumns_OneToOne_LeftJoin()
-        {
-            //Test merge by LEFT join of int and string columns
-
-            //Arrange
-            var left = new DataFrame();
-            left.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2 }));
-            left.Columns.Add(new Int32DataFrameColumn("G1", new[] { 1, 1, 2 }));
-            left.Columns.Add(new Int32DataFrameColumn("G2", new[] { 1, 2, 1 }));
-            left.Columns.Add(new StringDataFrameColumn("G3", new[] { "A", "B", "C" }));
-
-            var right = new DataFrame();
-            right.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2 }));
-            right.Columns.Add(new Int32DataFrameColumn("G1", new[] { 0, 1, 1 }));
-            right.Columns.Add(new Int32DataFrameColumn("G2", new[] { 1, 1, 2 }));
-            right.Columns.Add(new StringDataFrameColumn("G3", new[] { "Z", "Y", "B" }));
-
-            // Creates this case:
-            /*  -----------------------------
-             *      Left      |      Right
-             *   I  G1 G2 G3  |   I  G1 G2 G3
-             *  ------------------------------
-             *   0  1  1  A   |   0  0  1  Z
-             *   1  1  2  B   |   1  1  1  Y
-             *   2  2  1  C   |   2  1  2  B
-             */
-
-            /*
-             * Merge will result in a DataFrame like:
-             *   IL G1 G2 G3    IR              Merged:
-             *  -------------------------
-             *   0  1  1  A                      0 - N
-             *   1  1  2  B     2  1  2  B       1 - 2
-             *   2  2  1  C                      2 - N
-             */
-
-            //Act
-            var merge = left.Merge(right, new[] { "G1", "G2", "G3" }, new[] { "G1", "G2", "G3" });
-
-            //Assert
-            var expectedMerged = new (int? Left, int? Right)[] {
-                (0, null),
-                (1, 2),
-                (2, null)
-            };
-
-            Assert.Equal(expectedMerged.Length, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-
-            for (long i = 0; i < expectedMerged.Length; i++)
-            {
-                MatchRowsOnMergedDataFrame(merge, left, right, i, expectedMerged[i].Left, expectedMerged[i].Right);
-            }
-        }
-
-        [Fact]
-        public void TestMerge_ByThreeColumns_OneToOne_RightJoin()
-        {
-            //Test merge by RIGHT join of int and string columns
-
-            //Arrange
-            var left = new DataFrame();
-            left.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2 }));
-            left.Columns.Add(new Int32DataFrameColumn("G1", new[] { 1, 1, 2 }));
-            left.Columns.Add(new Int32DataFrameColumn("G2", new[] { 1, 2, 1 }));
-            left.Columns.Add(new StringDataFrameColumn("G3", new[] { "A", "B", "C" }));
-
-            var right = new DataFrame();
-            right.Columns.Add(new Int32DataFrameColumn("Index", new[] { 0, 1, 2 }));
-            right.Columns.Add(new Int32DataFrameColumn("G1", new[] { 0, 1, 1 }));
-            right.Columns.Add(new Int32DataFrameColumn("G2", new[] { 1, 1, 2 }));
-            right.Columns.Add(new StringDataFrameColumn("G3", new[] { "Z", "Y", "B" }));
-
-            // Creates this case:
-            /*  -----------------------------
-             *      Left      |      Right
-             *   I  G1 G2 G3  |   I  G1 G2 G3
-             *  ------------------------------
-             *   0  1  1  A   |   0  0  1  Z
-             *   1  1  2  B   |   1  1  1  Y
-             *   2  2  1  C   |   2  1  2  B
-             */
-
-            /*
-             * Merge will result in a DataFrame like:
-             *   IL G1 G2 G3    IR              Merged:
-             *  -------------------------
-             *                  0  0  1  Z       N - 0
-             *                  1  1  1  Y       N - 1
-             *   1  1  2  B     2  1  2  B       1 - 2
-             */
-
-            //Act
-            var merge = left.Merge(right, new[] { "G1", "G2", "G3" }, new[] { "G1", "G2", "G3" }, joinAlgorithm: JoinAlgorithm.Right);
-
-            //Assert
-            var expectedMerged = new (int? Left, int? Right)[] {
-                (null, 0),
-                (null, 1),
-                (1, 2)
-            };
-
-            Assert.Equal(expectedMerged.Length, merge.Rows.Count);
-            Assert.Equal(merge.Columns.Count, left.Columns.Count + right.Columns.Count);
-
-            for (long i = 0; i < expectedMerged.Length; i++)
-            {
-                MatchRowsOnMergedDataFrame(merge, left, right, i, expectedMerged[i].Left, expectedMerged[i].Right);
-            }
-        }
-
-        [Fact]
-        public void TestMerge_Issue5778()
-        {
-            DataFrame left = MakeDataFrameWithAllMutableColumnTypes(2, false);
-            DataFrame right = MakeDataFrameWithAllMutableColumnTypes(1);
-
-            DataFrame merge = left.Merge<int>(right, "Int", "Int");
-
-            Assert.Equal(2, merge.Rows.Count);
-            Assert.Equal(0, (int)merge.Columns["Int_left"][0]);
-            Assert.Equal(1, (int)merge.Columns["Int_left"][1]);
-            MatchRowsOnMergedDataFrame(merge, left, right, 0, 0, 0);
-            MatchRowsOnMergedDataFrame(merge, left, right, 1, 1, 0);
-        }
-
-        [Fact]
-        //Issue 6127
-        public void TestMerge_CorrectColumnTypes()
-        {
-            DataFrame left = MakeDataFrameWithAllMutableColumnTypes(2, false);
-            DataFrame right = MakeDataFrameWithAllMutableColumnTypes(1);
-
-            DataFrame merge = left.Merge<int>(right, "Int", "Int");
-
-            Assert.NotNull(merge.Columns.GetBooleanColumn("Bool_left"));
-            Assert.NotNull(merge.Columns.GetBooleanColumn("Bool_right"));
-
-            Assert.NotNull(merge.Columns.GetDecimalColumn("Decimal_left"));
-            Assert.NotNull(merge.Columns.GetDecimalColumn("Decimal_right"));
-
-            Assert.NotNull(merge.Columns.GetSingleColumn("Float_left"));
-            Assert.NotNull(merge.Columns.GetSingleColumn("Float_right"));
-
-            Assert.NotNull(merge.Columns.GetDoubleColumn("Double_left"));
-            Assert.NotNull(merge.Columns.GetDoubleColumn("Double_right"));
-
-            Assert.NotNull(merge.Columns.GetByteColumn("Byte_left"));
-            Assert.NotNull(merge.Columns.GetByteColumn("Byte_right"));
-
-            Assert.NotNull(merge.Columns.GetCharColumn("Char_left"));
-            Assert.NotNull(merge.Columns.GetCharColumn("Char_right"));
-
-            Assert.NotNull(merge.Columns.GetInt16Column("Short_left"));
-            Assert.NotNull(merge.Columns.GetInt16Column("Short_right"));
-
-            Assert.NotNull(merge.Columns.GetUInt16Column("Ushort_left"));
-            Assert.NotNull(merge.Columns.GetUInt16Column("Ushort_right"));
-
-            Assert.NotNull(merge.Columns.GetInt32Column("Int_left"));
-            Assert.NotNull(merge.Columns.GetInt32Column("Int_right"));
-
-            Assert.NotNull(merge.Columns.GetUInt32Column("Uint_left"));
-            Assert.NotNull(merge.Columns.GetUInt32Column("Uint_right"));
-
-            Assert.NotNull(merge.Columns.GetInt64Column("Long_left"));
-            Assert.NotNull(merge.Columns.GetInt64Column("Long_right"));
-
-            Assert.NotNull(merge.Columns.GetUInt64Column("Ulong_left"));
-            Assert.NotNull(merge.Columns.GetUInt64Column("Ulong_right"));
-
-            Assert.NotNull(merge.Columns.GetDateTimeColumn("DateTime_left"));
-            Assert.NotNull(merge.Columns.GetDateTimeColumn("DateTime_right"));
-
         }
 
         [Fact]
@@ -3467,87 +1224,6 @@ namespace Microsoft.Data.Analysis.Tests
         }
 
         [Fact]
-        public void TestBinaryOperationsOnExplodedNumericColumns()
-        {
-            DataFrame df = MakeDataFrameWithNumericAndBoolColumns(10, withNulls: false);
-            Int32DataFrameColumn ints = df.Columns["Int"] as Int32DataFrameColumn;
-            Int32DataFrameColumn res = ints.Add(1).Subtract(1).Multiply(10).Divide(10).LeftShift(2).RightShift(2);
-            Assert.True(res.ElementwiseEquals(ints).All());
-            Assert.True(res.ElementwiseGreaterThanOrEqual(ints).All());
-            Assert.True(res.ElementwiseLessThanOrEqual(ints).All());
-            Assert.False(res.ElementwiseNotEquals(ints).All());
-            Assert.False(res.ElementwiseGreaterThan(ints).All());
-            Assert.False(res.ElementwiseLessThan(ints).All());
-
-            // Test inPlace
-            Int32DataFrameColumn inPlace = ints.Add(1, inPlace: true).Subtract(1, inPlace: true).Multiply(10, inPlace: true).Divide(10, inPlace: true).LeftShift(2, inPlace: true).RightShift(2, inPlace: true).Add(100, inPlace: true);
-            Assert.True(inPlace.ElementwiseEquals(ints).All());
-            Assert.True(inPlace.ElementwiseGreaterThanOrEqual(ints).All());
-            Assert.True(inPlace.ElementwiseLessThanOrEqual(ints).All());
-            Assert.False(inPlace.ElementwiseNotEquals(ints).All());
-            Assert.False(inPlace.ElementwiseGreaterThan(ints).All());
-            Assert.False(inPlace.ElementwiseLessThan(ints).All());
-
-            Assert.False(inPlace.ElementwiseEquals(res).All());
-            Assert.True(inPlace.ElementwiseGreaterThanOrEqual(res).All());
-            Assert.False(inPlace.ElementwiseLessThanOrEqual(res).All());
-            Assert.True(inPlace.ElementwiseNotEquals(res).All());
-            Assert.True(inPlace.ElementwiseGreaterThan(res).All());
-            Assert.False(inPlace.ElementwiseLessThan(res).All());
-
-            // Test Bool column
-            BooleanDataFrameColumn bools = df.Columns["Bool"] as BooleanDataFrameColumn;
-            BooleanDataFrameColumn allFalse = bools.Or(true).And(true).Xor(true);
-            Assert.True(allFalse.ElementwiseEquals(false).All());
-
-            // Test inPlace
-            BooleanDataFrameColumn inPlaceAllFalse = bools.Or(true, inPlace: true).And(true, inPlace: true).Xor(true, inPlace: true);
-            Assert.True(inPlaceAllFalse.ElementwiseEquals(bools).All());
-
-            // Test Reverse Operations
-            Int32DataFrameColumn reverse = ints.ReverseAdd(1).ReverseSubtract(1).ReverseMultiply(-1);
-            Assert.True(reverse.ElementwiseEquals(ints).All());
-
-            // Test inPlace
-            Int32DataFrameColumn reverseInPlace = ints.ReverseAdd(1, inPlace: true).ReverseSubtract(1, inPlace: true).ReverseMultiply(-1, inPlace: true).ReverseDivide(100, inPlace: true);
-            Assert.True(reverseInPlace.ElementwiseEquals(ints).All());
-            Assert.False(reverseInPlace.ElementwiseEquals(reverse).All());
-        }
-
-        [Fact]
-        public void TestArrowStringApply()
-        {
-            ArrowStringDataFrameColumn column = CreateArrowStringColumn(10);
-            ArrowStringDataFrameColumn ret = column.Apply((string cur) =>
-            {
-                if (cur != null)
-                {
-                    return cur + "123";
-                }
-                return null;
-            });
-            for (long i = 0; i < column.Length; i++)
-            {
-                if (column[i] != null)
-                {
-                    Assert.Equal(column[i] + "123", ret[i]);
-                }
-                else
-                {
-                    Assert.Null(ret[i]);
-                }
-            }
-            Assert.Equal(1, ret.NullCount);
-
-            // Test null counts
-            ret = column.Apply((string cur) =>
-            {
-                return null;
-            });
-            Assert.Equal(column.Length, ret.NullCount);
-        }
-
-        [Fact]
         public void GetColumnTests()
         {
             DataFrame dataFrame = MakeDataFrameWithAllColumnTypes(10);
@@ -3628,40 +1304,6 @@ namespace Microsoft.Data.Analysis.Tests
         }
 
         [Fact]
-        public void Test_PrimitiveColumnNotEqualsNull()
-        {
-            var col = new DoubleDataFrameColumn("col", new double?[] { 1.23, null, 2, 3 });
-            var dfTest = new DataFrame(col);
-
-            var filteredNullDf = dfTest.Filter(dfTest["col"].ElementwiseNotEquals(null));
-
-            Assert.True(filteredNullDf.Columns.IndexOf("col") >= 0);
-            Assert.Equal(3, filteredNullDf.Columns["col"].Length);
-
-            Assert.Equal(1.23, filteredNullDf.Columns["col"][0]);
-            Assert.Equal(2.0, filteredNullDf.Columns["col"][1]);
-            Assert.Equal(3.0, filteredNullDf.Columns["col"][2]);
-        }
-
-        [Fact]
-        public void Test_PrimitiveColumnEqualsNull()
-        {
-            var index = new Int32DataFrameColumn("index", new int[] { 1, 2, 3, 4, 5 });
-            var col = new DoubleDataFrameColumn("col", new double?[] { 1.23, null, 2, 3, null }); ;
-            var dfTest = new DataFrame(index, col);
-
-            var filteredNullDf = dfTest.Filter(dfTest["col"].ElementwiseEquals(null));
-
-            Assert.True(filteredNullDf.Columns.IndexOf("col") >= 0);
-            Assert.True(filteredNullDf.Columns.IndexOf("index") >= 0);
-
-            Assert.Equal(2, filteredNullDf.Rows.Count);
-
-            Assert.Equal(2, filteredNullDf.Columns["index"][0]);
-            Assert.Equal(5, filteredNullDf.Columns["index"][1]);
-        }
-
-        [Fact]
         public void Test_StringColumnNotEqualsNull()
         {
             var col = new StringDataFrameColumn("col", new[] { "One", null, "Two", "Three" });
@@ -3693,111 +1335,6 @@ namespace Microsoft.Data.Analysis.Tests
 
             Assert.Equal(2, filteredNullDf.Columns["index"][0]);
             Assert.Equal(5, filteredNullDf.Columns["index"][1]);
-        }
-
-        [Fact]
-        public void Test_ArithmeticsAddWithNull()
-        {
-            // Arrange
-            //Number of elements shoult be higher than 8 to test SIMD
-            var left_column = new Int32DataFrameColumn("Left", new int?[] { 1, 1, null, null, 4, 5, 6, 7, 8, 9 });
-            var right_column = new Int32DataFrameColumn("Right", new int?[] { 1, null, 1, null, 4, 5, 6, 7, 8, 9 });
-
-            // Act
-            var sum = left_column + right_column;
-
-            // Assert
-            Assert.Equal(3, sum.NullCount);
-
-            Assert.Equal(2, sum[0]);  // 1 + 1
-            Assert.Null(sum[1]); // null + 1
-            Assert.Null(sum[2]); // 1 + null
-            Assert.Null(sum[3]); // null + null
-            Assert.Equal(8, sum[4]);
-            Assert.Equal(10, sum[5]);
-            Assert.Equal(12, sum[6]);
-            Assert.Equal(14, sum[7]);
-            Assert.Equal(16, sum[8]);
-            Assert.Equal(18, sum[9]);
-        }
-
-        [Fact]
-        public void Test_ArithmeticsAddScalarWithNull()
-        {
-            // Arrange
-            //Number of elements shoult be higher than 8 to test SIMD
-            var left_column = new Int32DataFrameColumn("Left", new int?[] { 0, 1, null, null, 4, 5, 6, 7, 8, null });
-
-            // Act
-            var sum = left_column + 5;
-
-            // Assert
-            Assert.Equal(3, sum.NullCount);
-
-            Assert.Equal(5, sum[0]);  // 1 + 1
-            Assert.Equal(6, sum[1]);  // 1 + 1
-            Assert.Null(sum[2]); // 1 + null
-            Assert.Null(sum[3]); // null + null
-            Assert.Equal(9, sum[4]);
-            Assert.Equal(10, sum[5]);
-            Assert.Equal(11, sum[6]);
-            Assert.Equal(12, sum[7]);
-            Assert.Equal(13, sum[8]);
-            Assert.Null(sum[9]);
-        }
-
-        [Fact]
-        public void Test_ArithmeticsDiffWithNull()
-        {
-            // Arrange
-            var left_column = new Int32DataFrameColumn("Left", new int?[] { 1, 1, null, null });
-            var right_column = new Int32DataFrameColumn("Right", new int?[] { 1, null, 1, null });
-
-            // Act
-            var diff = left_column - (right_column);
-
-            // Assert
-            Assert.Equal(3, diff.NullCount);
-            Assert.Equal(0, diff[0]);  // 1 - 1
-            Assert.Null(diff[1]); // null - 1
-            Assert.Null(diff[2]); // 1 - null
-            Assert.Null(diff[3]); // null - null
-        }
-
-        [Fact]
-        public void Test_ArithmeticsMultWithNull()
-        {
-            // Arrange
-            var left_column = new Int32DataFrameColumn("Left", new int?[] { 4, 1, null, null });
-            var right_column = new Int32DataFrameColumn("Right", new int?[] { 2, null, 1, null });
-
-            // Act
-            var mult = left_column * right_column;
-
-            // Assert
-            Assert.Equal(3, mult.NullCount);
-            Assert.Equal(8, mult[0]);  // 1 * 1
-            Assert.Null(mult[1]); // null * 1
-            Assert.Null(mult[2]); // 1 * null
-            Assert.Null(mult[3]); // null * null
-        }
-
-        [Fact]
-        public void Test_ArithmeticsDivWithNull()
-        {
-            // Arrange
-            var left_column = new Int32DataFrameColumn("Left", new int?[] { 4, 1, null, null });
-            var right_column = new Int32DataFrameColumn("Right", new int?[] { 2, null, 1, null });
-
-            // Act
-            var div = left_column / right_column;
-
-            // Assert
-            Assert.Equal(3, div.NullCount);
-            Assert.Equal(2, div[0]);  // 1 / 1
-            Assert.Null(div[1]); // null / 1
-            Assert.Null(div[2]); // 1 / null
-            Assert.Null(div[3]); // null / null
         }
     }
 }
